+++ CONTRIBUTING.rst
@@ -1,330 +1,249 @@
 .. highlight:: shell
 
+.. note::
+
+  Large portions of this document came from or are inspired by the `Xarray Contributing
+  Guide <https://docs.xarray.dev/en/stable/contributing.html>`_.
+
 ============
-Contributing
+Overview
 ============
 
-Contributions are welcome and greatly appreciated! Every little bit helps, and credit will always be given.
+xCDAT is a community-driven open source project and we welcome everyone who would like to
+contribute! Feel free to open a `GitHub Issue`_ or start a `GitHub Discussion`_ for bug
+reports, bug fixes, documentation improvements, enhancement suggestions, and other ideas.
 
-Types of Contributions
-----------------------
+We encourage discussion on any xCDAT related topic in the `GitHub Discussion`_ forum.
+You can also subscribe to our `mailing list`_ for news and announcements related to xCDAT,
+such as software version releases or future roadmap plans.
 
-xCDAT includes issue templates based on the contribution type: https://github.com/xCDAT/xcdat/issues/new/choose.
-Note, new contributions must be made under the Apache-2.0 with LLVM exception license.
+Please note that xCDAT has a `Code of Conduct`_. By participating in the xCDAT
+community, you agree to abide by its rules.
 
-Bug Report
-~~~~~~~~~~
+.. _GitHub Issue: https://github.com/xCDAT/xcdat/issues
+.. _GitHub Discussion: https://github.com/xCDAT/xcdat/discussions
+.. _Code of Conduct: CODE-OF-CONDUCT.rst
+.. _mailing list: https://groups.google.com/g/xcdat
 
-Look through the `GitHub Issues`_ for bugs to fix. Any unassigned issues tagged with "Type: Bug" is open for implementation.
+Where to start?
+===============
 
-Feature Request
-~~~~~~~~~~~~~~~
+If you are brand new to xCDAT or open-source development, we recommend going
+through the `GitHub Issues`_ page to find issues that interest you. If you are
+interested in opening a GitHub issue, you can use the templates `here <https://github.com/xCDAT/xcdat/issues/new/choose>`_
 
-Look through the `GitHub Issues`_ for feature suggestions. Any unassigned issues tagged with "Type: Enhancement" is open for implementation.
+Some issues are particularly suited for new contributors by the label
+`Documentation <https://github.com/xCDAT/xcdat/labels/type%3A%20documentation>`_ and
+`good first issue <https://github.com/xCDAT/xcdat/labels/good-first-issue>`_ where
+you could start out. These are well documented issues, that do not require a deep
+understanding of the internals of xCDAT. Once you've found an interesting issue, you can
+return here to get your development environment setup.
 
-If you are proposing a feature:
+**New contributions must be made under the Apache-2.0 with LLVM exception license.**
 
-* Explain in detail how it would work.
-* Keep the scope as narrow as possible, to make it easier to implement.
-* Remember that this is a open-source project, and that contributions are welcome :)
+Documentation Updates
+~~~~~~~~~~~~~~~~~~~~~
 
-Features must meet the following criteria before they are considered for implementation:
+Contributing to the `documentation <https://xcdat.readthedocs.io/en/latest/>`_ is an
+excellent way to help xCDAT. You don't need to be an expert in xCDAT to help with
+documentation!
 
-1. Feature is not implemented by ``xarray``
-2. Feature is not implemented in another actively developed xarray-based package
+Bug Reports and enhancement requests
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+Bug reports are an important part of making xCDAT more stable. Having a complete bug
+report will allow others to reproduce the bug and provide insight into fixing.
 
[…]
