+++ CONTRIBUTING.rst
@@ -10,8 +10,8 @@ Overview
 ============
 
 xCDAT is a community-driven open source project and we welcome everyone who would like to
-contribute! Feel free to open a `GitHub Issue`_ or start a `GitHub Discussion`_ for bug
-reports, bug fixes, documentation improvements, enhancement suggestions, and other ideas.
+contribute! Feel free to open a `GitHub Issue`_ for bug reports, bug fixes,
+documentation improvements, enhancement suggestions, and other ideas.
 
 We encourage discussion on any xCDAT related topic in the `GitHub Discussion`_ forum.
 You can also subscribe to our `mailing list`_ for news and announcements related to xCDAT,
@@ -20,21 +20,16 @@ such as software version releases or future roadmap plans.
 Please note that xCDAT has a `Code of Conduct`_. By participating in the xCDAT
 community, you agree to abide by its rules.
 
-.. _GitHub Issue: https://github.com/xCDAT/xcdat/issues
-.. _GitHub Discussion: https://github.com/xCDAT/xcdat/discussions
-.. _Code of Conduct: CODE-OF-CONDUCT.rst
-.. _mailing list: https://groups.google.com/g/xcdat
-
 Where to start?
-===============
+---------------
 
 If you are brand new to xCDAT or open-source development, we recommend going
 through the `GitHub Issues`_ page to find issues that interest you. If you are
-interested in opening a GitHub issue, you can use the templates `here <https://github.com/xCDAT/xcdat/issues/new/choose>`_
+interested in opening a GitHub Issue, you can use these `templates`_.
 
 Some issues are particularly suited for new contributors by the label
-`Documentation <https://github.com/xCDAT/xcdat/labels/type%3A%20documentation>`_ and
-`good first issue <https://github.com/xCDAT/xcdat/labels/good-first-issue>`_ where
+`type: docs <https://github.com/xCDAT/xcdat/labels/type%3A%20docs>`_ and
+`good-first-issue <https://github.com/xCDAT/xcdat/labels/good-first-issue>`_ where
 you could start out. These are well documented issues, that do not require a deep
 understanding of the internals of xCDAT. Once you've found an interesting issue, you can
 return here to get your development environment setup.
@@ -44,12 +39,14 @@ return here to get your development environment setup.
 Documentation Updates
 ~~~~~~~~~~~~~~~~~~~~~
 
-Contributing to the `documentation <https://xcdat.readthedocs.io/en/latest/>`_ is an
-excellent way to help xCDAT. You don't need to be an expert in xCDAT to help with
-documentation!
+Contributing to the `documentation`_ is an excellent way to help xCDAT. Anything from
+fixing typos to improving sentence structure or API docstrings/examples can go a long
+way in making xCDAT more user-friendly.
 
-Bug Reports and enhancement requests
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+To open a documentation update request, please use the `documentation update template`_.
+
+Bug reports
+~~~~~~~~~~~
 Bug reports are an important part of making xCDAT more stable. Having a complete bug
 report will allow others to reproduce the bug and provide insight into fixing.
 
@@ -57,26 +54,46 @@ Trying out the bug-producing code on the ``main`` branch is often a worthwhile e
 to confirm that the bug still exists. It is also worth searching existing bug reports and
 pull requests to see if the issue has already been reported and/or fixed.
 
-.. _GitHub Issues: https://github.com/xCDAT/xcdat/issues
+To open a new bug report, please use this issue `bug report template`_.
+
+Enhancement requests
+~~~~~~~~~~~~~~~~~~~~
+Enhancements are a great way to improve xCDAT's capabilities for the broader scientific
+community.
+
[…]
