### 2024-02-04 4fa2158
Initial work on #596
