+++ xcdat/spatial.py
@@ -29,7 +29,7 @@ from xcdat.utils import _if_multidim_dask_array_then_load
 #: Type alias for a dictionary of axis keys mapped to their bounds.
 AxisWeights = Dict[Hashable, xr.DataArray]
 #: Type alias for supported spatial axis keys.
-SpatialAxis = Literal["X", "Y"]
+SpatialAxis = Literal["X", "Y", "Z"]
 SPATIAL_AXES: Tuple[SpatialAxis, ...] = get_args(SpatialAxis)
 #: Type alias for a tuple of floats/ints for the regional selection bounds.
 RegionAxisBounds = Tuple[float, float]
@@ -73,10 +73,12 @@ class SpatialAccessor:
         keep_weights: bool = False,
         lat_bounds: Optional[RegionAxisBounds] = None,
         lon_bounds: Optional[RegionAxisBounds] = None,
+        lev_bounds: Optional[RegionAxisBounds] = None,
     ) -> xr.Dataset:
         """
-        Calculates the spatial average for a rectilinear grid over an optionally
-        specified regional domain.
+        Calculates the weighted spatial and/or vertical average for a
+        rectilinear grid over an optionally specified regional and/or vertical
+        domain.
 
         Operations include:
 
@@ -101,7 +103,7 @@ class SpatialAccessor:
             average.
         axis : List[SpatialAxis]
             List of axis dimensions to average over, by default ["X", "Y"].
-            Valid axis keys include "X" and "Y".
+            Valid axis keys include "X", "Y", and "Z".
         weights : {"generate", xr.DataArray}, optional
             If "generate", then weights are generated. Otherwise, pass a
             DataArray containing the regional weights used for weighted
@@ -122,6 +124,10 @@ class SpatialAccessor:
             ignored if ``weights`` are supplied. The lower bound can be larger
             than the upper bound (e.g., across the prime meridian, dateline), by
             default None.
+        lev_bounds : Optional[RegionAxisBounds], optional
+            A tuple of floats/ints for the regional lower and upper level
+            boundaries. This arg is used when calculating axis weights, but is
+            ignored if ``weights`` are supplied. The default is None.
 
         Returns
         -------
@@ -143,11 +149,15 @@ class SpatialAccessor:
         >>>
         >>> ds.lon.attrs["axis"]
         >>> X
+        >>>
+        >>> ds.level.attrs["axis"]
+        >>> Z
 
         Set the 'axis' attribute for the required coordinates if it isn't:
 
         >>> ds.lat.attrs["axis"] = "Y"
         >>> ds.lon.attrs["axis"] = "X"
+        >>> ds.level.attrs["axis"] = "Z"
 
         Call spatial averaging method:
 
@@ -167,6 +177,10 @@ class SpatialAccessor:
 
         >>> ts_zonal = ds.spatial.average("tas", axis=["X"])["tas"]
 
+        Get the vertical average (between 100 and 1000 hPa):
+
+        >>> ta_column = ds.spatial.average("ta", axis=["Z"], lev_bounds=(100, 1000))["ta"]
+
         Using custom weights for averaging:
 
         >>> # The shape of the weights must align with the data var.
@@ -178,6 +192,12 @@ class SpatialAccessor:
         >>>
         >>> ts_global = ds.spatial.average("tas", axis=["X", "Y"],
         >>>     weights=weights)["tas"]
+
+        Notes:
+        ------
+        Weights are generally computed as the difference between the bounds. If
+        sub-selecting a region, the units must match the axis units (e.g.,
+        Pa/hPa or m/km).
         """
         ds = self._dataset.copy()
         dv = _get_data_var(ds, data_var)
@@ -188,7 +208,11 @@ class SpatialAccessor:
[…]
