### 2024-02-06 9e407e3
Update temporal averaging, bounds, and weights docs - Add more specific information on how temporal averaging grouping is performed - Add information on assumed time frequencies for `get_time_bounds()` - Add info about how weights are not properly assigned if one time point spans across the time interval being averaged into
