### 2024-02-08 9be64c3
augmenting axis fall-back table (#584)
