### 2024-02-14 98ea3d7
add unit tests for vertical averaging
