### 2024-02-16 5af9ae5
Apply suggestions from code review
Co-authored-by: Stephen Po-Chedley <pochedley@gmail.com>
