### 2024-02-16 fd36dae
Remove unnecessary space after docstring bullet points
