+++ xcdat/temporal.py
@@ -175,7 +175,6 @@ class TemporalAccessor:
         ----------
         data_var: str
             The key of the data variable for calculating averages
-
         weighted : bool, optional
             Calculate averages using weights, by default True.
 
@@ -193,6 +192,7 @@ class TemporalAccessor:
             (e.g., a timepoint on 2020-06-01 with bounds that begin in May 2020
             and end in June 2020), the weights will not be assigned properly.
             See explanation in the Notes section below.
+
         keep_weights : bool, optional
             If calculating averages using weights, keep the weights in the
             final dataset output, by default False.
@@ -260,7 +260,6 @@ class TemporalAccessor:
             * "month": groups by (year, month) for monthly averages.
             * "day": groups by (year, month, day) for daily averages.
             * "hour": groups by (year, month, day, hour) for hourly averages.
-
         weighted : bool, optional
             Calculate averages using weights, by default True.
 
@@ -278,6 +277,7 @@ class TemporalAccessor:
             (e.g., a timepoint on 2020-06-01 with bounds that begin in May 2020
             and end in June 2020), the weights will not be assigned properly.
             See explanation in the Notes section below.
+
         keep_weights : bool, optional
             If calculating averages using weights, keep the weights in the
             final dataset output, by default False.
@@ -430,7 +430,6 @@ class TemporalAccessor:
               present) are dropped to avoid inconsistencies when calculating
               climatologies. Refer to [1]_ for more details on this
               implementation decision.
-
         weighted : bool, optional
             Calculate averages using weights, by default True.
 
@@ -606,7 +605,6 @@ class TemporalAccessor:
         ----------
         data_var: str
             The key of the data variable for calculating departures.
-
         freq : Frequency
             The frequency of time to group by.
 
@@ -618,7 +616,6 @@ class TemporalAccessor:
               present) are dropped to avoid inconsistencies when calculating
               climatologies. Refer to [2]_ for more details on this
               implementation decision.
-
         weighted : bool, optional
             Calculate averages using weights, by default True.
 
