+++ xcdat/dataset.py
@@ -80,7 +80,6 @@ def open_dataset(
           of the coordinates. If desired, refer to
           :py:func:`xarray.Dataset.bounds.add_time_bounds` if you require more
           granular configuration for how "T" bounds are generated.
-
     decode_times: bool, optional
         If True, attempt to decode times encoded in the standard NetCDF
         datetime format into cftime.datetime objects. Otherwise, leave them
@@ -95,11 +94,10 @@ def open_dataset(
         Either `(-180, 180)` or `(0, 360)`, by default None. Supported options
         include:
 
-          * None:  use the current orientation (if the longitude axis exists)
-          * (-180, 180): represents [-180, 180) in math notation
-          * (0, 360): represents [0, 360) in math notation
-
-    kwargs : Dict[str, Any]
+        * None:  use the current orientation (if the longitude axis exists)
+        * (-180, 180): represents [-180, 180) in math notation
+        * (0, 360): represents [0, 360) in math notation
+    **kwargs : Dict[str, Any]
         Additional arguments passed on to ``xarray.open_dataset``. Refer to the
         [1]_ xarray docs for accepted keyword arguments.
 
@@ -155,25 +153,24 @@ def open_mfdataset(
         Paths to dataset files. Paths can be given as strings or as pathlib.Path
         objects. Supported options include:
 
-          * Directory path (e.g., ``"path/to/files"``), which is converted
-            to a string glob of `*.nc` files
-          * String glob (e.g., ``"path/to/files/*.nc"``), which is expanded
-            to a 1-dimensional list of file paths
-          * File path to dataset (e.g., ``"path/to/files/file1.nc"``)
-          * List of file paths (e.g., ``["path/to/files/file1.nc", ...]``).
-            If concatenation along more than one dimension is desired, then
-            ``paths`` must be a nested list-of-lists (see [2]_
-            ``xarray.combine_nested`` for details).
-          * File path to an XML file with a ``directory`` attribute (e.g.,
-            ``"path/to/files"``). If ``directory`` is set to a blank string
-            (""), then the current directory is substituted ("."). This option
-            is intended to support the CDAT CDML dialect of XML files, but it
-            can work with any XML file that has the ``directory`` attribute.
-            Refer to [4]_ for more information on CDML. NOTE: This feature is
-            deprecated in v0.6.0 and will be removed in the subsequent release.
-            CDAT (including cdms2/CDML) is in maintenance only mode and marked
-            for end-of-life by the end of 2023.
-
+        * Directory path (e.g., ``"path/to/files"``), which is converted
+          to a string glob of `*.nc` files
+        * String glob (e.g., ``"path/to/files/*.nc"``), which is expanded
+          to a 1-dimensional list of file paths
+        * File path to dataset (e.g., ``"path/to/files/file1.nc"``)
+        * List of file paths (e.g., ``["path/to/files/file1.nc", ...]``).
+          If concatenation along more than one dimension is desired, then
+          ``paths`` must be a nested list-of-lists (see [2]_
+          ``xarray.combine_nested`` for details).
+        * File path to an XML file with a ``directory`` attribute (e.g.,
+          ``"path/to/files"``). If ``directory`` is set to a blank string
+          (""), then the current directory is substituted ("."). This option
+          is intended to support the CDAT CDML dialect of XML files, but it
+          can work with any XML file that has the ``directory`` attribute.
+          Refer to [4]_ for more information on CDML. NOTE: This feature is
+          deprecated in v0.6.0 and will be removed in the subsequent release.
+          CDAT (including cdms2/CDML) is in maintenance only mode and marked
+          for end-of-life by the end of 2023.
     add_bounds: List[CFAxisKey] | None | bool
         List of CF axes to try to add bounds for (if missing), by default
[…]
