### 2024-02-21 78dedf0
[PR]: Update documentation on temporal averaging, usage of bounds, and generation of weights (#601)
Co-authored-by: Stephen Po-Chedley <pochedley@gmail.com>
