### 2024-02-23 96eaae2
Fixes docstrings and adds comments
