+++ xcdat/regridder/regrid2.py
@@ -119,11 +119,13 @@ def _regrid(
     other_sizes = list(other_dims.values())
 
     data_shape = [y_length * x_length] + other_sizes
+    # output data is always float32 in original code
     output_data = np.zeros(data_shape, dtype=np.float32)
 
     is_2d = input_data_var.ndim <= 2
 
-    # need to optimize
+    # TODO: need to optimize further, investigate using ufuncs and dask arrays
+    # TODO: how common is lon by lat data? may need to reshape
     for y in range(y_length):
         y_seg = np.take(input_data, lat_mapping[y], axis=y_index)
 
@@ -134,6 +136,10 @@ def _regrid(
 
             output_seg_index = y * x_length + x
 
+            # using the `out` argument is more performant, places data directly into
+            # array memory rather than allocating a new variable. wasn't working for
+            # single element output, needs further investigation as we may not need
+            # branch
             if is_2d:
                 output_data[output_seg_index] = np.divide(
                     np.sum(
@@ -208,9 +214,9 @@ def _map_latitude(src: np.ndarray, dst: np.ndarray) -> Tuple[List, List]:
     Parameters
     ----------
     src : np.ndarray
-        DataArray containing the source latitude bounds.
+        Array containing the source latitude bounds.
     dst : np.ndarray
-        DataArray containing the destination latitude bounds.
+        Array containing the destination latitude bounds.
 
     Returns
     -------
@@ -222,16 +228,23 @@ def _map_latitude(src: np.ndarray, dst: np.ndarray) -> Tuple[List, List]:
 
     dst_length = dst_south.shape[0]
 
+    # finds contributing source cells for each destination cell based on bounds values
+    # output is a list of lists containing the contributing cell indexes
+    # e.g. let src_south be [90, 45, 0, -45], source_north be [45, 0, -45, -90],
+    # dst_north[x] be 70, and dst_south[x] be -70 then the result would be [[1, 2]]
     mapping = [
         np.where(np.logical_and(src_south < dst_north[x], src_north > dst_south[x]))[0]
         for x in range(dst_length)
     ]
 
+    # finds minimum and maximum bounds for each output cell, considers source and
+    # destination bounds for each cell
     bounds = [
         (np.minimum(dst_north[x], src_north[y]), np.maximum(dst_south[x], src_south[y]))
         for x, y in enumerate(mapping)
     ]
 
+    # convert latitude to cell weight (difference of length from equator)
     weights = [
         (np.sin(np.deg2rad(x)) - np.sin(np.deg2rad(y))).reshape((-1, 1))
         for x, y in bounds
@@ -247,9 +260,9 @@ def _map_longitude(src: np.ndarray, dst: np.ndarray) -> Tuple[List, List]:
     Parameters
     ----------
     src : np.ndarray
-        DataArray containing source longitude bounds.
+        Array containing source longitude bounds.
     dst : np.ndarray
-        DataArray containing destination longitude bounds.
+        Array containing destination longitude bounds.
 
     Returns
     -------
@@ -259,6 +272,7 @@ def _map_longitude(src: np.ndarray, dst: np.ndarray) -> Tuple[List, List]:
     src_west, src_east = _extract_bounds(src)
     dst_west, dst_east = _extract_bounds(dst)
 
+    # align source and destination longitude
     shifted_src_west, shifted_src_east, shift = _align_axis(
         src_west,
         src_east,
@@ -268,6 +282,8 @@ def _map_longitude(src: np.ndarray, dst: np.ndarray) -> Tuple[List, List]:
     src_length = src_west.shape[0]
     dst_length = dst_west.shape[0]
 
+    # finds contributing source cells for each destination cell based on bounds values
+    # output is a list of lists containing the contributing cell indexes
     mapping = [
         np.where(
[…]
