### 2024-02-27 1dc3866
Update notebook with scheduler info
