### 2024-02-28 #6426 @editorialbot
[REVIEW]: xCDAT: A Python package for simple climate data analysis on structured grids
**Submitting author:** @tomvothecoder (<a href="http://orcid.org/0000-0002-2461-0191">Tom Vo</a>)
**Repository:** https://github.com/xCDAT/xcdat
**Branch with paper.md** (empty if default branch): main
**Version:** v0.7.1
