### 2024-03-15 f120c2d
Add list of projects using xCDAT (#617)
- Update GitHub issue templates to use latest tags
- Update release drafter labels
