+++ .github/ISSUE_TEMPLATE/enhancement_request.yml
@@ -1,7 +1,7 @@
-name: Feature Request
-description: Suggest an idea for xCDAT
-title: "[Feature]: "
-labels: ["Type: Enhancement"]
+name: Enhancement Request
+description: Suggest an enhancement for xCDAT
+title: "[Enhancement]: "
+labels: ["type: enhancement"]
 assignees: []
 body:
     - type: textarea
@@ -21,7 +21,7 @@ body:
           description: |
               A clear and concise description of what you want to happen.
       validations:
-        required: false
+          required: false
 
     - type: textarea
       id: alternatives
