### 2024-03-18 xCDAT/xcdat#619 @mgrover1
[Bug]: Unable to create conda environment for development environment on M1 Mac
### What happened?

I tried creating the conda environment for this project (https://github.com/xCDAT/xcdat/blob/main/conda-env/dev.yml) on an M1 macbook (arm architecture) and ran into issues with the pandoc installation. This is also related to

- https://github.com/xCDAT/xcdat-validation/issues/51
- https://github.com/openjournals/joss-reviews/issues/6426

### What did you expect to happen? Are there are possible answers you came across?

I expected to be able to run

```bash
conda env create -f conda-env/dev.yml

```

but ran into the following issue

```bash
Could not solve for environment specs
The following package could not be installed
└─ pandoc 3.1.1**  does not exist (perhaps a typo or a missing channel).
```

### Minimal Complete Verifiable Example (MVCE)

```python
conda env create -f conda-env/dev.yml
```

### Relevant log output

```python
[+] 0.1s
conda-forge/noarch                                            No change
pkgs/main/noarch                                              No change
pkgs/r/osx-arm64                                              No change
pkgs/r/noarch                                                 No change
pkgs/main/osx-arm64                                @  11.1MB/s  0.2s
conda-forge/osx-arm64                                8.9MB @  17.8MB/s  0.5s

Looking for: ["python[version='>=3.9']", "cf_xarray[version='>=0.7.3']", 'cftime', 'dask', 'lxml', 'netcdf4', "numpy[version='>=1.23.0']", 'pandas', 'python-dateutil', "xarray[version='>=2022.02.0']", "xesmf[version='>=0.7.0']", 'xgcm', "matplotlib-base[version='>=3.7.0']", 'nc-time-axis=1.4.1', 'sphinx=5.3.0', 'sphinx-autosummary-accessors=2022.4.0', 'sphinx-book-theme=1.0.1', 'sphinx-copybutton=0.5.1', 'nbsphinx=0.9.1', 'pandoc=3.1.1', 'ipython=8.11.0', 'gsw-xarray=0.3.0', 'types-python-dateutil', 'black=23.3.0', 'flake8=6.0.0', 'flake8-isort=6.0.0', 'isort=5.12.0', 'mypy=1.4.0', 'pre-commit=3.2.0', 'pytest', 'pytest-cov', 'jupyterlab', 'tbump']

Could not solve for environment specs
The following package could not be installed
└─ pandoc 3.1.1**  does not exist (perhaps a typo or a missing channel).
(base) mgrover@evswl129 xcdat % Could not solve for environment specs
The following package could not be installed
└─ pandoc 3.1.1**  does not exist (perhaps a typo or a missing channel).
```

### Anything else we need to know?

It should be possible to build a development environment on a recent Mac machine.

### Environment

Unable to do this here as the environment will not build.
