### 2024-03-18 #6426 @mgrover1
@tomvo - I went through with a first cut. I am struggling to reproduce the benchmarks you mention due to trying to run on an M1 machine, I plan to test with my intel Mac later this week.

Most of the comments are tracked in the associated the issues, the only other comment I have is that your reference in line 150 should change from

```
https://doi.org/https://doi.org/10.1029/2022MS003156
```

to

```
https://doi.org/10.1029/2022MS003156
```

The writing is well done, and the documentation is fantastic - just a couple of environment issues that likely would be fixed with some additional CI/cross architecture tests.
