### 2024-03-18 xCDAT/xcdat#620 @pochedls
Thank you for your review! We will fix the documentation here (and [ensure general accuracy](https://github.com/xCDAT/xcdat/issues/589)). The daily dataset does not have bounds, which can be added on open (or after open with `ds_hourly = ds_hourly.bounds.add_missing_bounds(["T"])`):

```python
ds_hourly = xcdat.open_dataset(filepath2, chunks={"time": "auto"}, add_bounds=["T"])
```

xCDAT used to add the bounds automatically, but we decided that a user should make this decision.
