### 2024-03-18 xCDAT/xcdat#619 @pochedls
@tomvothecoder – I was able to reproduce this on an M2 Mac. Do you think we can drop the `pandoc` dependency for `xcdat-dev` (maybe we could have a separate `maintainer` env for repo administration/documentation and `xcdat-dev` could be purely for contributing to xcdat functionality)?
