### 2024-03-18 xCDAT/xcdat#621 @mgrover1
BUG: Incorrect DOI Reference in Paper Draft
### What happened?

One of the DOI links is improperly formatted in the paper draft - https://github.com/xCDAT/xcdat/blob/f995ec6716a69d2613b08dcd3e5de0bed20ba122/docs/paper/paper.bib#L79

This is related to the review
https://github.com/openjournals/joss-reviews/issues/6426

### What did you expect to happen? Are there are possible answers you came across?

This should just be the DOI ID portion (10.1029/2022MS003156), but instead includes the URL as well. Removing the prefix should fix this.

### Minimal Complete Verifiable Example (MVCE)

_No response_

### Relevant log output

_No response_

### Anything else we need to know?

_No response_

### Environment

N/A
