### 2024-03-18 0f16e90
Remove pins for doc dependencies - Fixes issue where `pandoc=3.11` does not support osx-arm64
