### 2024-03-18 xCDAT/xcdat#619 @tomvothecoder
> @tomvothecoder – I was able to reproduce this on an M2 Mac. Do you think we can drop the `pandoc` dependency for `xcdat-dev` (maybe we could have a separate `maintainer` env for repo administration/documentation and `xcdat-dev` could be purely for contributing to xcdat functionality)?

I was also able to reproduce this issue on my M2 Mac, but resolved it in #622 by dropping the pins for documentation dependencies. `pandoc>=3.12` supports `osx-arm64`. Can you check out this PR branch and try it out too?

We should keep `pandoc` and other doc dependencies in the dev env so there is single env for dev, which is easier to use and maintain compared to multiple dev envs (e.g., one for code dev, one for doc).
