+++ docs/paper/paper.bib
@@ -76,7 +76,7 @@
    number     = {12},
    pages      = {e2022MS003156},
    keywords   = {DOE E3SM, climate modeling},
-   doi        = {https://doi.org/10.1029/2022MS003156},
+   doi        = {10.1029/2022MS003156},
    url        = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2022MS003156},
    eprint     = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2022MS003156},
    note       = {e2022MS003156 2022MS003156},
