### 2024-03-18 xCDAT/xcdat#619 @tomvothecoder
@pochedls Thanks for confirming. I merged PR #622 to fix this issue.

@mgrover1 Can you try to building the dev env on your M1 Mac using the latest `main` dev env?
