### 2024-03-18 xCDAT/xcdat#621 @mgrover1
@tomvothecoder - looks like this commit resolves the issue
https://github.com/xCDAT/xcdat/pull/567/commits/0d5a98131e8f06f7e0339d0c846ece6856b410f6

I think we can close this, as this issue refers to the PR not necessarily the main branch?
