### 2024-03-18 xCDAT/xcdat#620 @tomvothecoder
This is now resolved in #623
