### 2024-03-18 xCDAT/xcdat#620 @tomvothecoder
I also checked if this issue was present in the other temporal averaging notebook, but it only uses the monthly file that already has bounds.
