### 2024-03-18 3584504
Doc updates - Add sphinx_design for grid tables - Update homepage for simple navigation - Add `overview.rst` page to move overview content - Add `getting-start-guides/` with sub-nesting of pages for getting started - Add `team.rst` with grid of team members - Update `gallery.rst` with grid of examples - Update `conf.py` configuration for `sphinx_design` grid, custom CSS stylesheet, and custom xarray functions for generating grids - Update `.gitignore`
