+++ docs/getting-started-guide/index.rst
@@ -1,0 +1,15 @@
+################
+Getting Started
+################
+
+The getting started guide aims to get you using xarray productively as quickly as possible.
+It is designed as an entry point for new users, and it provided an introduction to xarray's main concepts.
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   Project Overview <overview>
+   Installation <installation>
+   xCDAT on HPC / Jupyter <getting-started-hpc-jupyter>
+   faqs
