+++ docs/team.rst
@@ -1,0 +1,14 @@
+Team
+-----
+
+Current core developers
+~~~~~~~~~~~~~~~~~~~~~~~
+
+xCDAT core developers are responsible for the ongoing organizational maintenance and technical direction of the xCDAT project.
+
+The current core developers team comprises:
+
+.. include:: team-panel.txt
+
+
+The full list of contributors is on our `GitHub Contributors Page <https://github.com/xCDAT/xcdat/graphs/contributors>`__.
