+++ conda-env/dev.yml
@@ -24,14 +24,15 @@ dependencies:
     - nc-time-axis=1.4.1
     # Documentation
     # ==================
-    - sphinx=5.3.0
-    - sphinx-autosummary-accessors=2022.4.0
-    - sphinx-book-theme=1.0.1
-    - sphinx-copybutton=0.5.1
-    - nbsphinx=0.9.1
-    - pandoc=3.1.1
-    - ipython=8.11.0 # Required for nbsphinx syntax highlighting
-    - gsw-xarray=0.3.0 # Required for vertical regridding example
+    - sphinx
+    - sphinx-autosummary-accessors
+    - sphinx-book-theme
+    - sphinx-copybutton
+    - nbsphinx
+    - sphinx-design
+    - pandoc
+    - ipython # Required for nbsphinx syntax highlighting
+    - gsw-xarray # Required for vertical regridding example
     # Quality Assurance
     # ==================
     - types-python-dateutil

[…]
