### 2024-03-18 f2368d4
Fix xarray ref
