+++ docs/getting-started-guide/index.rst
@@ -2,8 +2,7 @@
 Getting Started
 ################
 
-The getting started guide aims to get you using xarray productively as quickly as possible.
-It is designed as an entry point for new users, and it provided an introduction to xarray's main concepts.
+The getting started guide aims to get you using xCDAT productively as quickly as possible.
 
 .. toctree::
    :maxdepth: 2
