+++ docs/demos.rst
@@ -3,10 +3,12 @@ Presentations and Demos
 
 This page includes relevant xCDAT presentations, demos, and papers.
 
+.. include:: demos.txt
+
 .. toctree::
     :maxdepth: 2
     :hidden:
 
     AMS 2023 Abstract <https://ams.confex.com/ams/103ANNUAL/meetingapp.cgi/Paper/412648>
     demos/1-25-23-cwss-seminar/introduction-to-xcdat.ipynb
-    A Gentle Introduction to xCDAT (Xarray Climate Data Analysis Tools) - Video introduction <https://youtu.be/sJpQ9vKDKa8?feature=shared>
+    ESGF2 - A Gentle Introduction to xCDAT (video introduction) <https://youtu.be/sJpQ9vKDKa8?feature=shared>
