### 2024-03-20 b86f0ce
Update documentation styling for easier navigation (#624)
