+++ docs/getting-started-guide/index.rst
@@ -1,0 +1,14 @@
+################
+Getting Started
+################
+
+The getting started guide aims to get you using xCDAT productively as quickly as possible.
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   Project Overview <overview>
+   Installation <installation>
+   xCDAT on HPC / Jupyter <getting-started-hpc-jupyter>
+   faqs
