+++ docs/index.rst
@@ -1,224 +1,65 @@
-xCDAT: Xarray Climate Data Analysis Tools
-=========================================
+xCDAT Documentation
+====================
 
-xCDAT is an extension of `xarray`_ for climate data analysis on structured grids. It
-serves as a modern successor to the Community Data Analysis Tools (`CDAT`_) library.
+xCDAT (Xarray Climate Data Analysis Tools) is an extension of `xarray`_ for climate data analysis on structured grids. It serves as a modern successor to the Community Data Analysis Tools (`CDAT`_) library.
+
+.. _xarray: https://github.com/pydata/xarray
+.. _CDAT: https://github.com/CDAT/cdat
 
 **Useful links**:
-`Documentation <https://xcdat.readthedocs.io>`__ |
 `Code Repository <https://github.com/xCDAT/xcdat>`__ |
 `Issues <https://github.com/xCDAT/xcdat/issues>`__ |
 `Discussions <https://github.com/xCDAT/xcdat/discussions>`__ |
 `Releases <https://github.com/xCDAT/xcdat/releases>`__ |
 `Mailing List <https://groups.google.com/g/xcdat>`__
 
-Project Motivation
-------------------
+.. grid:: 1 1 2 2
+    :gutter: 2
 
-The goal of xCDAT is to provide generalizable features and utilities for simple and
-robust analysis of climate data. xCDAT's design philosophy is focused on reducing the
-overhead required to accomplish certain tasks in xarray. xCDAT aims to be compatible
-with structured grids that are `CF-compliant`_ (e.g., CMIP6). Some key xCDAT features
-are inspired by or ported from the core CDAT library, while others leverage powerful
-libraries in the xarray ecosystem (e.g., `xESMF`_, `xgcm`_, `cf_xarray`_) to deliver
-robust APIs.
+    .. grid-item-card:: Getting started
+        :img-top: _static/icons/play.svg
+        :link: getting-started-guide/index
+        :link-type: doc
 
-The xCDAT core team's mission is to provide a maintainable and extensible package
-that serves the needs of the climate community in the long-term. We are excited
-to be working on this project and hope to have you onboard!
+        New to xCDAT? Check out the getting started guides.
 
-.. _CF-compliant: https://cfconventions.org/
-.. _xarray: https://github.com/pydata/xarray
-.. _CDAT: https://github.com/CDAT/cdat
+    .. grid-item-card::  Examples
+        :img-top: _static/icons/plot.svg
+        :link: gallery
+        :link-type: doc
 
-Getting Started
----------------
+        This gallery demonstrates how to use some of the features in xCDAT through
+        Jupyter Notebooks with real-world examples.
 
-The best resource for getting started is the `xCDAT documentation website`_.
-Our documentation provides general guidance for setting up xCDAT in an Anaconda
-environment on your local `computer`_ or on an `HPC/Jupyter`_ environment. We also
-include an `API Overview`_ and `Gallery`_ to highlight xCDAT functionality.
+    .. grid-item-card::  API reference
+        :img-top: _static/icons/wrench.svg
+        :link: api
+        :link-type: doc
 
-.. _xCDAT documentation website: https://xcdat.readthedocs.io/en/stable/
-.. _computer: https://xcdat.readthedocs.io/en/stable/getting-started.html
-.. _HPC/Jupyter: https://xcdat.readthedocs.io/en/stable/getting-started-hpc-jupyter.html
-.. _API Overview: https://xcdat.readthedocs.io/en/stable/api.html
-.. _Gallery: https://xcdat.readthedocs.io/en/stable/gallery.html
+        The reference guide contains a detailed description of the xCDAT API.
[…]
