+++ AUTHORS.rst
@@ -1,9 +1,0 @@
-========
-The Team
-========
-
-* Tom Vo <tomvothecoder@gmail.com>
-* Jason Boutte
-* Stephen Po-Chedley
-* Jill Chengzhu Zhang
-* Jiwoo Lee
