### 2024-03-20 #6426 @mgrover1
## Round 2 of reviews

Thanks for addressing all of the previous comments!

@tomvothecoder - is there anyway to make the datasets mentioned in the benchmarks more accessible? I understand one of the files is 105 GB, but I think it would help with reproducibility if there were someway to download these locally and execute the validation scripts you all put together.

see https://github.com/xCDAT/xcdat-validation/issues/52
