+++ xcdat/dataset.py
@@ -46,17 +46,13 @@ Paths = Union[
 def open_dataset(
     path: str | os.PathLike[Any] | BufferedIOBase | AbstractDataStore,
     data_var: Optional[str] = None,
-    add_bounds: List[CFAxisKey] | None | bool = ["X", "Y"],
+    add_bounds: List[CFAxisKey] | None = ["X", "Y"],
     decode_times: bool = True,
     center_times: bool = False,
     lon_orient: Optional[Tuple[float, float]] = None,
     **kwargs: Dict[str, Any],
 ) -> xr.Dataset:
     """Wraps ``xarray.open_dataset()`` with post-processing options.
-
-    .. deprecated:: v0.6.0
-        ``add_bounds`` boolean arguments (True/False) are being deprecated.
-        Please use either a list (e.g., ["X", "Y"]) to specify axes or ``None``.
 
     Parameters
     ----------
@@ -133,7 +129,7 @@ def open_dataset(
 def open_mfdataset(
     paths: str | NestedSequence[str | os.PathLike],
     data_var: Optional[str] = None,
-    add_bounds: List[CFAxisKey] | None | Literal[False] = ["X", "Y"],
+    add_bounds: List[CFAxisKey] | None = ["X", "Y"],
     decode_times: bool = True,
     center_times: bool = False,
     lon_orient: Optional[Tuple[float, float]] = None,
@@ -142,10 +138,6 @@ def open_mfdataset(
     **kwargs: Dict[str, Any],
 ) -> xr.Dataset:
     """Wraps ``xarray.open_mfdataset()`` with post-processing options.
-
-    .. deprecated:: v0.6.0
-        ``add_bounds`` boolean arguments (True/False) are being deprecated.
-        Please use either a list (e.g., ["X", "Y"]) to specify axes or ``None``.
 
     Parameters
     ----------
@@ -240,15 +232,6 @@ def open_mfdataset(
     in-memory copy you are manipulating in xarray is modified: the original file
     on disk is never touched.
 
-    The CDAT "Climate Data Markup Language" (CDML) is a deprecated dialect of
-    XML with a defined set of attributes. CDML is still used by current and
-    former users of CDAT. To enable CDML users to adopt xCDAT more easily in
-    their workflows, xCDAT can parse XML/CDML files for the ``directory``
-    to generate a glob or list of file paths. Refer to [4]_ for more information
-    on CDML. NOTE: This feature is deprecated in v0.6.0 and will be removed in
-    the subsequent release. CDAT (including cdms2/CDML) is in maintenance only
-    mode and marked for end-of-life by the end of 2023.
-
     References
     ----------
     .. [2] https://docs.xarray.dev/en/stable/generated/xarray.combine_nested.html
@@ -258,13 +241,6 @@ def open_mfdataset(
     if isinstance(paths, str) or isinstance(paths, pathlib.Path):
         if os.path.isdir(paths):
             paths = _parse_dir_for_nc_glob(paths)
-        elif _is_xml_filepath(paths):
-            warnings.warn(
-                "`open_mfdataset()` will no longer support CDML/XML paths after "
-                "v0.6.0 because CDAT is marked for end-of-life at the end of 2023.",
-                DeprecationWarning,
-            )
-            paths = _parse_xml_for_nc_glob(paths)
 
     preprocess = partial(_preprocess, decode_times=decode_times, callable=preprocess)
 
@@ -422,58 +398,6 @@ def decode_time(dataset: xr.Dataset) -> xr.Dataset:
     return ds
 
 
-def _is_xml_filepath(paths: str | pathlib.Path) -> bool:
-    """Checks if the ``paths`` argument is a path to an XML file.
-
-    Parameters
-    ----------
-    paths : str | pathlib.Path
[…]
