+++ xcdat/regridder/accessor.py
@@ -117,124 +117,6 @@ class RegridderAccessor:
 
         return coord_var, bounds_var
 
-    # TODO Either provide generic `horizontal` and `vertical` methods or tool specific
-    def horizontal_xesmf(
-        self,
-        data_var: str,
-        output_grid: xr.Dataset,
-        **options: Any,
-    ) -> xr.Dataset:
-        """
-        Deprecated, will be removed with 0.7.0 release.
-
-        Extends the xESMF library for horizontal regridding between structured
-        rectilinear and curvilinear grids.
-
-        This method extends ``xESMF`` by automatically constructing the
-        ``xe.XESMFRegridder`` object, preserving source bounds, and generating
-        missing bounds. It regrids ``data_var`` in the dataset to
-        ``output_grid``.
-
-        Option documentation :py:func:`xcdat.regridder.xesmf.XESMFRegridder`
-
-        Parameters
-        ----------
-        data_var: str
-            Name of the variable in the `xr.Dataset` to regrid.
-        output_grid : xr.Dataset
-            Dataset containing output grid.
-        options : Dict[str, Any]
-            Dictionary with extra parameters for the regridder.
-
-        Returns
-        -------
-        xr.Dataset
-            With the ``data_var`` variable on the grid defined in ``output_grid``.
-
-        Raises
-        ------
-        ValueError
-            If tool is not supported.
-
-        Examples
-        --------
-
-        Generate output grid:
-
-        >>> output_grid = xcdat.create_gaussian_grid(32)
-
-        Regrid data to output grid using xesmf:
-
-        >>> ds.regridder.horizontal_xesmf("ts", output_grid)
-        """
-        warnings.warn(
-            "`horizontal_xesmf` will be deprecated in 0.7.x, please migrate to using "
-            "`horizontal(..., tool='xesmf')` method.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-
-        regridder = HORIZONTAL_REGRID_TOOLS["xesmf"](self._ds, output_grid, **options)
-
-        return regridder.horizontal(data_var, self._ds)
-
-    # TODO Either provide generic `horizontal` and `vertical` methods or tool specific
-    def horizontal_regrid2(
-        self,
-        data_var: str,
-        output_grid: xr.Dataset,
-        **options: Any,
-    ) -> xr.Dataset:
-        """
-        Deprecated, will be removed with 0.7.0 release.
-
-        Pure python implementation of CDAT's regrid2 horizontal regridder.
-
-        Regrids ``data_var`` in dataset to ``output_grid`` using regrid2's
-        algorithm.
-
-        Options documentation :py:func:`xcdat.regridder.regrid2.Regrid2Regridder`
-
-        Parameters
-        ----------
-        data_var: str
-            Name of the variable in the `xr.Dataset` to regrid.
-        output_grid : xr.Dataset
-            Dataset containing output grid.
-        options : Dict[str, Any]
-            Dictionary with extra parameters for the regridder.
-
-        Returns
-        -------
-        xr.Dataset
-            With the ``data_var`` variable on the grid defined in ``output_grid``.
-
-        Raises
-        ------
-        ValueError
-            If tool is not supported.
-
-        Examples
-        --------
-        Generate output grid:
-
-        >>> output_grid = xcdat.create_gaussian_grid(32)
-
-        Regrid data to output grid using regrid2:
-
-        >>> ds.regridder.horizontal_regrid2("ts", output_grid)
-        """
-        warnings.warn(
-            "`horizontal_regrid2` will be deprecated in 0.7.x, please migrate to using "
-            "`horizontal(..., tool='regrid2')` method.",
-            DeprecationWarning,
[…]
