+++ xcdat/regridder/grid.py
@@ -454,13 +454,8 @@ def create_grid(
         ]
     ] = None,
     attrs: Optional[Dict[str, str]] = None,
-    **kwargs: CoordOptionalBnds,
 ) -> xr.Dataset:
     """Creates a grid dataset using the specified axes.
-
-    .. deprecated:: v0.6.0
-        ``**kwargs`` argument is being deprecated, please migrate to
-        ``x``, ``y``, or ``z`` arguments to create future grids.
 
     Parameters
     ----------
@@ -515,17 +510,6 @@ def create_grid(
     >>> )
     >>> grid = create_grid(z=z)
     """
-    if np.all([item is None for item in (x, y, z)]) and len(kwargs) == 0:
-        raise ValueError("Must pass at least 1 axis to create a grid.")
-    elif np.all([item is None for item in (x, y, z)]) and len(kwargs) > 0:
-        warnings.warn(
-            "**kwargs will be deprecated, see docstring and use 'x', 'y', or 'z' arguments",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-
-        return _deprecated_create_grid(**kwargs)
-
     axes = {"x": x, "y": y, "z": z}
     ds = xr.Dataset(attrs={} if attrs is None else attrs.copy())
 
@@ -555,46 +539,6 @@ def create_grid(
         ds = ds.assign_coords({coords.name: coords})
 
     return ds
-
-
-def _deprecated_create_grid(**kwargs: CoordOptionalBnds) -> xr.Dataset:
-    coords = {}
-    data_vars = {}
-
-    for name, data in kwargs.items():
-        if name in VAR_NAME_MAP["X"]:
-            coord, bnds = _prepare_coordinate(name, data, **COORD_DEFAULT_ATTRS["X"])
-        elif name in VAR_NAME_MAP["Y"]:
-            coord, bnds = _prepare_coordinate(name, data, **COORD_DEFAULT_ATTRS["Y"])
-        elif name in VAR_NAME_MAP["Z"]:
-            coord, bnds = _prepare_coordinate(name, data, **COORD_DEFAULT_ATTRS["Z"])
-        else:
-            raise ValueError(
-                f"Coordinate {name} is not valid, reference "
-                "`xcdat.axis.VAR_NAME_MAP` for valid options."
-            )
-
-        coords[name] = coord
-
-        if bnds is not None:
-            bnds = bnds.copy()
-
-            if isinstance(bnds, np.ndarray):
-                bnds = xr.DataArray(
-                    name=f"{name}_bnds",
-                    data=bnds.copy(),
-                    dims=[name, "bnds"],
-                )
-
-            data_vars[bnds.name] = bnds
-
-            coord.attrs["bounds"] = bnds.name
-
-    grid = xr.Dataset(data_vars, coords=coords)
-
-    grid = grid.bounds.add_missing_bounds(axes=["X", "Y"])
-
-    return grid
 
 
 def _prepare_coordinate(name: str, data: CoordOptionalBnds, **attrs: Any):
