### 2024-03-20 bf34e25
Remove tests for deprecated code - Remove `lxml` dependency
