+++ tests/test_regrid.py
@@ -891,73 +891,6 @@ class TestGrid:
         ):
             grid.create_grid(x=(self.lon, self.lon_bnds, self.lat))  # type: ignore[arg-type]
 
-    def test_deprecated_unexpected_coordinate(self):
-        lev = np.linspace(1000, 1, 2)
-
-        with pytest.raises(
-            ValueError,
-            match="Coordinate mass is not valid, reference `xcdat.axis.VAR_NAME_MAP` for valid options.",
-        ):
-            grid.create_grid(lev=lev, mass=np.linspace(10, 20, 2))
-
-    def test_deprecated_create_grid_lev(self):
-        lev = np.linspace(1000, 1, 2)
-        lev_bnds = np.array([[1499.5, 500.5], [500.5, -498.5]])
-
-        with warnings.catch_warnings(record=True) as w:
-            new_grid = grid.create_grid(lev=(lev, lev_bnds))
-
-            assert len(w) == 1
-            assert issubclass(w[0].category, DeprecationWarning)
-            assert (
-                str(w[0].message)
-                == "**kwargs will be deprecated, see docstring and use 'x', 'y', or 'z' arguments"
-            )
-
-        assert np.array_equal(new_grid.lev, lev)
-        assert np.array_equal(new_grid.lev_bnds, lev_bnds)
-
-    def test_deprecated_create_grid(self):
-        lat = np.array([-45, 0, 45])
-        lon = np.array([30, 60, 90, 120, 150])
-        lat_bnds = np.array([[-67.5, -22.5], [-22.5, 22.5], [22.5, 67.5]])
-        lon_bnds = np.array([[15, 45], [45, 75], [75, 105], [105, 135], [135, 165]])
-
-        new_grid = grid.create_grid(lat=lat, lon=lon)
-
-        assert np.array_equal(new_grid.lat, lat)
-        assert np.array_equal(new_grid.lat_bnds, lat_bnds)
-        assert new_grid.lat.units == "degrees_north"
-        assert np.array_equal(new_grid.lon, lon)
-        assert np.array_equal(new_grid.lon_bnds, lon_bnds)
-        assert new_grid.lon.units == "degrees_east"
-
-        da_lat = xr.DataArray(
-            name="lat",
-            data=lat,
-            dims=["lat"],
-            attrs={"units": "degrees_north", "axis": "Y"},
-        )
-        da_lon = xr.DataArray(
-            name="lon",
-            data=lon,
-            dims=["lon"],
-            attrs={"units": "degrees_east", "axis": "X"},
-        )
-        da_lat_bnds = xr.DataArray(name="lat_bnds", data=lat_bnds, dims=["lat", "bnds"])
-        da_lon_bnds = xr.DataArray(name="lon_bnds", data=lon_bnds, dims=["lon", "bnds"])
-
-        new_grid = grid.create_grid(
-            lat=(da_lat, da_lat_bnds), lon=(da_lon, da_lon_bnds)
-        )
-
-        assert np.array_equal(new_grid.lat, lat)
-        assert np.array_equal(new_grid.lat_bnds, lat_bnds)
-        assert new_grid.lat.units == "degrees_north"
-        assert np.array_equal(new_grid.lon, lon)
-        assert np.array_equal(new_grid.lon_bnds, lon_bnds)
-        assert new_grid.lon.units == "degrees_east"
-
     def test_uniform_grid(self):
         new_grid = grid.create_uniform_grid(-90, 90, 4.0, -180, 180, 5.0)
 
