+++ tests/test_dataset.py
@@ -324,48 +324,6 @@ class TestOpenDataset:
 
         assert result.identical(expected)
 
-    def test_raises_deprecation_warning_when_passing_add_bounds_true(self):
-        ds_no_bounds = generate_dataset(
-            decode_times=True, cf_compliant=True, has_bounds=False
-        )
-        ds_no_bounds.to_netcdf(self.file_path)
-
-        with warnings.catch_warnings(record=True) as w:
-            result = open_dataset(self.file_path, add_bounds=True)
-
-            assert len(w) == 1
-            assert issubclass(w[0].category, DeprecationWarning)
-            assert str(w[0].message) == (
-                "`add_bounds=True` will be deprecated after v0.6.0. Please use a list "
-                "of axis strings instead (e.g., `add_bounds=['X', 'Y']`)."
-            )
-
-        expected = generate_dataset(
-            decode_times=True, cf_compliant=True, has_bounds=True
-        )
-        expected = expected.drop_vars("time_bnds")
-        del expected["time"].attrs["bounds"]
-
-        assert result.identical(expected)
-
-    def test_raises_deprecation_warning_when_passing_add_bounds_false(self):
-        ds_no_bounds = generate_dataset(
-            decode_times=True, cf_compliant=True, has_bounds=False
-        )
-        ds_no_bounds.to_netcdf(self.file_path)
-
-        with warnings.catch_warnings(record=True) as w:
-            result = open_dataset(self.file_path, add_bounds=False)
-
-            assert len(w) == 1
-            assert issubclass(w[0].category, DeprecationWarning)
-            assert str(w[0].message) == (
-                "`add_bounds=False` will be deprecated after v0.6.0. Please use "
-                "`add_bounds=None` instead."
-            )
-
-        assert result.identical(ds_no_bounds)
-
 
 class TestOpenMfDataset:
     @pytest.fixture(autouse=True)
@@ -415,74 +373,6 @@ class TestOpenMfDataset:
 
         result = open_mfdataset(self.file_path1, add_bounds=None)
         assert result.identical(ds)
-
-    def test_raises_error_if_xml_does_not_have_root_directory_attr(self):
-        ds1 = generate_dataset(decode_times=False, cf_compliant=False, has_bounds=True)
-        ds1.to_netcdf(self.file_path1)
-        ds2 = generate_dataset(decode_times=False, cf_compliant=False, has_bounds=True)
-        ds2 = ds2.rename_vars({"ts": "tas"})
-        ds2.to_netcdf(self.file_path2)
-
-        # Create the XML file
-        xml_path = f"{self.dir}/datasets.xml"
-        page = etree.Element("dataset")
-        doc = etree.ElementTree(page)
-        doc.write(xml_path, xml_declaration=True, encoding="utf-16")
-
-        with pytest.raises(KeyError):
-            open_mfdataset(xml_path, decode_times=True)
-
-    def test_opens_datasets_from_xml_using_str_path(self):
-        ds1 = generate_dataset(decode_times=False, cf_compliant=False, has_bounds=True)
-        ds1.to_netcdf(self.file_path1)
-        ds2 = generate_dataset(decode_times=False, cf_compliant=False, has_bounds=True)
-        ds2 = ds2.rename_vars({"ts": "tas"})
-        ds2.to_netcdf(self.file_path2)
-
-        # Create the XML file
-        xml_path = f"{self.dir}/datasets.xml"
-        page = etree.Element("dataset", directory=str(self.dir))
-        doc = etree.ElementTree(page)
-        doc.write(xml_path, xml_declaration=True, encoding="utf-16")
-
[…]
