+++ tests/test_regrid.py
@@ -1,7 +1,6 @@
 import datetime
 import re
 import sys
-import warnings
 from unittest import mock
 
 import numpy as np
