### 2024-03-20 f273ddc
Remove deprecated `_prepare_coordinate()` - Add ValueError conditional back to `create_grid()`
