### 2024-03-20 232b0ab
Fix tests and update docstring for `create_grid()`
