### 2024-03-20 f50c158
Fix tests and update docstring for `create_grid()`
