### 2024-03-20 90fadf0
Fix `test_global_mean_grid` and `test_zonal_grid`
