### 2024-03-21 5de4d7f
[PR]: Revert augmenting axis fall-back table for Z axis (#629)
