### 2024-03-21 424bd11
Fix type annotation for Python 3.9
