+++ xcdat/regridder/grid.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from typing import Dict, List, Optional, Tuple, Union
 
 import numpy as np
