### 2024-03-21 #6426 @tomvothecoder
> @tomvothecoder - is there anyway to make the datasets mentioned in the benchmarks more accessible? I understand one of the files is 105 GB, but I think it would help with reproducibility if there were someway to download these locally and execute the validation scripts you all put together.

@mgrover1 Good point! I'll check to see if those datasets are available on ESGF and Globus. I'll also update the instructions for running the performance benchmark script to make it easier to reproduce the results.
