### 2024-03-28 85d1c7c
Fix regrid2 `open_mfdataset()` support with Dask Arrays - Load dask arrays into memory using `.values`
