### 2024-03-28 7e900a0
Remove unneeded dask conversion logic
