### 2024-04-02 472c04b
[PR]: Remove deprecated features and APIs for next release (#628)
* Remove deprecated features and APIs
- Remove `horizontal_xesmf()` and  `horizontal_regrid2()`
- Remove `**kwargs` from `create_grid()` and `_deprecated_create_grid()`
- Remove `add_bounds` accepting boolean arg in `open_dataset()`, `open_mfdataset()` and `_postprocess_dataset()`
* Remove tests for deprecated code
- Remove `lxml` dependency
* Remove `_prepare_coordinate()`
* Fix `test_global_mean_grid` and `test_zonal_grid`
