+++ conda-env/dev.yml
@@ -10,7 +10,6 @@ dependencies:
     - cf_xarray >=0.7.3 # Constrained because https://github.com/xarray-contrib/cf-xarray/issues/467
     - cftime
     - dask
-    - lxml # TODO: Remove this in v0.7.0 once cdml/XML support is dropped
     - netcdf4
     - numpy >=1.23.0 # This version of numpy includes support for Python 3.11.
     - pandas
