### 2024-04-02 a88da42
Add Steve suggestion to replace mamba with conda
