### 2024-04-02 9ac1189
Replace mamba with conda in installation guide
