### 2024-04-03 5aeee56
Fixes applying mask from input_grid
