### 2024-04-03 7af50a0
Fixes replacing 1e20 with nan after regridding
