### 2024-04-03 36b3d00
exclude nan or alternative missing values from horizontal regridding process
