### 2024-04-04 dfc3ebc
Fixes applying mask
