### 2024-04-04 6719f03
Adds omitted for cells that have no weight
