### 2024-04-04 ff57de4
Add deprecation warning for "drop_incomplete_djf"
