### 2024-04-04 a18ab83
Add deprecation warning for "drop_incomplete_djf"
