+++ tests/test_temporal.py
@@ -1,4 +1,5 @@
 import logging
+import warnings
 
 import cftime
 import numpy as np
@@ -620,7 +621,7 @@ class TestGroupAverage:
             },
         )
 
-        assert result.identical(expected)
+        xr.testing.assert_identical(result, expected)
 
     def test_weighted_seasonal_averages_with_DJF_and_drop_incomplete_seasons(self):
         ds = generate_dataset(decode_times=True, cf_compliant=True, has_bounds=True)
@@ -872,7 +873,7 @@ class TestGroupAverage:
             },
         )
 
-        assert result.identical(expected)
+        xr.testing.assert_identical(result, expected)
 
     def test_weighted_custom_seasonal_averages_with_seasons_spanning_calendar_years(
         self,
@@ -1151,7 +1152,7 @@ class TestClimatology:
             ds.temporal.climatology(
                 "ts",
                 "season",
-                season_config={"dec_mode": "DJF", "drop_incomplete_djf": True},
+                season_config={"dec_mode": "DJF", "drop_incomplete_seasons": True},
                 reference_period=("01-01-2000", "01-01-2000"),
             )
 
@@ -1159,7 +1160,7 @@ class TestClimatology:
             ds.temporal.climatology(
                 "ts",
                 "season",
-                season_config={"dec_mode": "DJF", "drop_incomplete_djf": True},
+                season_config={"dec_mode": "DJF", "drop_incomplete_seasons": True},
                 reference_period=("01-01-2000"),
             )
 
@@ -1169,7 +1170,7 @@ class TestClimatology:
         result = ds.temporal.climatology(
             "ts",
             "season",
-            season_config={"dec_mode": "DJF", "drop_incomplete_djf": True},
+            season_config={"dec_mode": "DJF", "drop_incomplete_seasons": True},
             reference_period=("2000-01-01", "2000-06-01"),
         )
 
@@ -1201,7 +1202,7 @@ class TestClimatology:
                 "freq": "season",
                 "weighted": "True",
                 "dec_mode": "DJF",
-                "drop_incomplete_djf": "True",
+                "drop_incomplete_seasons": "True",
             },
         )
 
@@ -1215,6 +1216,70 @@ class TestClimatology:
             "season",
             season_config={"dec_mode": "DJF", "drop_incomplete_seasons": True},
         )
+
+        expected = ds.copy()
+        expected = expected.drop_dims("time")
+        expected_time = xr.DataArray(
+            data=np.array(
+                [
+                    cftime.DatetimeGregorian(1, 1, 1),
+                    cftime.DatetimeGregorian(1, 4, 1),
+                    cftime.DatetimeGregorian(1, 7, 1),
+                    cftime.DatetimeGregorian(1, 10, 1),
+                ],
+            ),
+            coords={
+                "time": np.array(
+                    [
+                        cftime.DatetimeGregorian(1, 1, 1),
+                        cftime.DatetimeGregorian(1, 4, 1),
+                        cftime.DatetimeGregorian(1, 7, 1),
+                        cftime.DatetimeGregorian(1, 10, 1),
+                    ],
+                ),
+            },
+            attrs={
+                "axis": "T",
+                "long_name": "time",
+                "standard_name": "time",
+                "bounds": "time_bnds",
+            },
+        )
+        expected["ts"] = xr.DataArray(
+            name="ts",
+            data=np.ones((4, 4, 4)),
+            coords={"lat": expected.lat, "lon": expected.lon, "time": expected_time},
+            dims=["time", "lat", "lon"],
+            attrs={
+                "operation": "temporal_avg",
+                "mode": "climatology",
+                "freq": "season",
+                "weighted": "True",
[…]
