### 2024-04-04 561709f
Add test to cover missing line in `_drop_incomplete_season`
