+++ .github/workflows/build_workflow.yml
@@ -107,10 +107,11 @@ jobs:
                   pytest
 
             - name: Upload Coverage Report
-              uses: codecov/codecov-action@v3
+              uses: codecov/codecov-action@v4
               with:
                   file: "tests_coverage_reports/coverage.xml"
                   fail_ci_if_error: true
+                  token: ${{ secrets.CODECOV_TOKEN }}
 
     # `build-result` is a workaround to skipped matrix jobs in `build` not being considered "successful",
     # which can block PR merges if matrix jobs are required status checks.
