### 2024-04-10 173bba2
Update names in changelog to GitHub hyperlink
