### 2024-04-10 65c0be6
Improve readability of documentation (#632)
Co-authored-by: Tom Vo <tomvothecoder@gmail.com>
