+++ tbump.toml
@@ -2,7 +2,7 @@
 github_url = "https://github.com/xCDAT/xcdat"
 
 [version]
-current = "0.6.1"
+current = "0.7.0"
 
 # Example of a semver regexp.
 # Make sure this matches current_version before
