+++ docs/getting-started-guide/installation.rst
@@ -95,19 +95,19 @@ To update to a specific version of ``xcdat``:
    >>> mamba activate <ENV_NAME>
    >>> mamba update xcdat=<MAJOR.MINOR.PATCH>
 
-Jupyter Users set `ESMFMKFILE` env variable
--------------------------------------------
+Jupyter Users set ``ESMFMKFILE`` env variable
+---------------------------------------------
 
-If you are a Jupyter user, the `ESMFMKFILE` environment variable will need to be set
+If you are a Jupyter user, the ``ESMFMKFILE`` environment variable will need to be set
 either directly on the machine or through your Jupyter Notebook.
 
-This env variable is normally set when calling `conda activate` with the conda
-environment that has `xesmf`. However, Jupyter does not run `conda activate` when using
-the Python kernel associated with the environment so `ESMFMKFILE` is not set, resulting 
-in `ImportError: The ESMFMKFILE environment variable is not available.` (related [GitHub
+This env variable is normally set when calling ``conda activate`` with the conda
+environment that has ``xesmf``. However, Jupyter does not run ``conda activate`` when using
+the Python kernel associated with the environment so ``ESMFMKFILE`` is not set, resulting 
+in ``ImportError: The ESMFMKFILE environment variable is not available.`` (related [GitHub
 Issue](https://github.com/xCDAT/xcdat/issues/574)).
 
-To set the `EMSFMKFILE` in a Jupyter Notebook add:
+To set the ``EMSFMKFILE`` in a Jupyter Notebook add:
 
 .. code-block:: python
 
