### 2024-04-30 cf96e8b
Fix dask logo ref link
