### 2024-04-30 dda5f5b
Fix dask logo ref link
