+++ docs/getting-started-guide/faqs.rst
@@ -7,17 +7,15 @@ Metadata Interpretation
 
 What types of datasets does ``xcdat`` primarily focus on?
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-``xcdat`` supports datasets with structured grids that follow the `CF convention`_, but
-will also strive to support datasets with common non-CF compliant metadata (e.g., time
-units in "months since ..." or "years since ...").
+``xcdat`` supports datasets with structured grids that follow the `CF convention`_.
 
 .. _CF convention: http://cfconventions.org/
 
-What structured grids does ``xcdat`` support?
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+What structured grids are supported by  ``xcdat``?
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 xCDAT aims to be a generalizable package that is compatible with structured grids that
-are **CF-compliant** (e.g., CMIP6). xCDAT’s horizontal regridder supports grids that are
-supported by Regrid2 and xESMF  (curvilinear and rectilinear).
+are **CF-compliant** (e.g., CMIP6). xCDAT's spatial averager currently supports
+rectilinear grids, and the horizontal regridder curvilinear and rectilinear grids.
 
 How does ``xcdat`` interpret dataset metadata?
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -25,6 +23,9 @@ How does ``xcdat`` interpret dataset metadata?
 ``xcdat`` methods and functions usually accept an ``axis`` argument (e.g.,
 ``ds.temporal.average("ts")``). This argument is internally mapped to ``cf_xarray``
 mapping tables that interpret the CF attributes.
+
+xCDAT also includes its own "fall-back" mapping table that maps axes to their commonly accepted names
+(e.g., "X" maps to "longitude" and "lon").
 
 .. _cf_xarray: https://cf-xarray.readthedocs.io/en/latest/index.html
 
@@ -60,17 +61,30 @@ Handling Bounds
 
 How are bounds generated in xCDAT?
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-xCDAT generates bounds by using coordinate points as the midpoint between their lower
-and upper bounds.
+For the X, Y, and Z axes, xCDAT generates bounds by using coordinate points as the
+midpoint between their lower and upper bounds.
+
+For the T axis, xCDAT can generate bounds either by 1) time frequency (default method)
+or 2) midpoints.
+
+   1. time frequency: create time bounds as the start and end of each timestep's period
+      using either the inferred or specified time frequency.
+   2. midpoint: create time bounds using time coordinates as the midpoint between their
+      upper and lower bounds.
+
+For more information, visit the documentation for these APIs:
+- https://xcdat.readthedocs.io/en/stable/generated/xarray.Dataset.bounds.add_missing_bounds.html
+- https://xcdat.readthedocs.io/en/stable/generated/xarray.Dataset.bounds.add_time_bounds.html
+- https://xcdat.readthedocs.io/en/stable/generated/xarray.Dataset.bounds.add_bounds.html
 
 Does xCDAT support generating bounds for multiple axis coordinate systems in the same dataset?
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
[…]
