### 2024-04-30 b1c2c81
Fix missing word in `faqs.rst`
