### 2024-05-07 #6426 @brian-rose
@tomvothecoder this is a minor annoyance, but the author for the reference to Dask is showing up in the rendered paper as "Team, D. D.". Is there some way to override the parsing of [the bib file entry for Dask](https://github.com/xCDAT/xcdat/blob/0d5a98131e8f06f7e0339d0c846ece6856b410f6/docs/paper/paper.bib#L231) so that the citation becomes more readable?
