+++ docs/paper/paper.bib
@@ -228,7 +228,7 @@
 
 @manual{dask:2016,
    title      = {Dask: Library for dynamic task scheduling},
-   author     = {Dask Development Team},
+   author     = {Dask-Development-Team},
    year       = {2016},
    url        = {https://dask.org},
 }
