### 2024-05-10 #6426 @tomvothecoder
> > @tomvothecoder this is a minor annoyance, but the author for the reference to Dask is showing up in the rendered paper as "Team, D. D.". Is there some way to override the parsing of [the bib file entry for Dask](https://github.com/xCDAT/xcdat/blob/0d5a98131e8f06f7e0339d0c846ece6856b410f6/docs/paper/paper.bib#L231) so that the citation becomes more readable?
>
> I'll try to see if there is a way to make that citation render in a readable format (maybe I'll try dashes). I used the same Dask citation from the [xclim paper](https://joss.theoj.org/papers/10.21105/joss.05415) and it also pops up as "Team, D. D." on there.

I updated the Dask author citation to "Dask-Development-Team"  [here](https://github.com/xCDAT/xcdat/pull/567/commits/a5b005a62a83abd47cb34932ab9d96d8abb4a1b3).
