### 2024-05-14 #6426 @brian-rose
I've encountered some issues that have so far prevented me from reproducing the benchmark calculation that appears in the article.

I've tried to follow the steps outlined [on this page](https://github.com/xCDAT/xcdat-validation/blob/main/scripts/performance-benchmarks/README.md).

First, I was unable to create the necessary mamba environment on my Mac laptop. I commented on that here: https://github.com/xCDAT/xcdat-validation/issues/51

I got a bit farther on a linux server. But I encountered an error running one of the python scripts, which I documented here: https://github.com/xCDAT/xcdat-validation/issues/56
