### 2024-05-15 #6426 @tomvothecoder
> I've encountered some issues that have so far prevented me from reproducing the benchmark calculation that appears in the article.
>
> I've tried to follow the steps outlined [on this page](https://github.com/xCDAT/xcdat-validation/blob/main/scripts/performance-benchmarks/README.md?rgh-link-date=2024-05-14T20%3A48%3A00Z).
>
> First, I was unable to create the necessary mamba environment on my Mac laptop. I commented on that here: [xCDAT/xcdat-validation#51](https://github.com/xCDAT/xcdat-validation/issues/51)
>
> I got a bit farther on a linux server. But I encountered an error running one of the python scripts, which I documented here: [xCDAT/xcdat-validation#56](https://github.com/xCDAT/xcdat-validation/issues/56)

Thanks for reporting these issues @brian-rose. I'll take a look at these today and will aim to find solutions ASAP!
