### 2024-05-24 #6426 @brian-rose
@arfon a quick update on my review: my attempt to reproduce the benchmark calculation in the manuscript took me on a long detour.

With help from the author team (https://github.com/xCDAT/xcdat-validation/issues/56) I made some progress but ultimately concluded that I cannot reproduce the calculation because of the very large memory requirements of the benchmark script. With help from my IT support, I ran the script on the largest memory machine that I have access to on my campus (768 GB) but it was not enough.

So, I will not be able to check off the "Reproducibility" box. I've suggested that the authors update the documentation for their benchmark scripts to clarify the large memory requirement. In my opinion, that should be sufficient for publication.

I will finally finish working through functionality and documentation to complete my review.
