### 2024-05-31 f3ef9df
add skipna parameter
