### 2024-05-31 23d018d
add skipna parameter
