### 2024-05-31 24951ec
add skipna parameter
