### 2024-05-31 696fb21
add skipna parameter
