### 2024-05-31 ad5e9e4
add skipna parameter
