### 2024-05-31 225cac9
add skipna parameter
