### 2024-05-31 8204188
add skipna where it was missed
