### 2024-05-31 534ad4b
add skipna where it was missed
