### 2024-05-31 b2763d1
add skipna where it was missed
