### 2024-06-05 7c25b7c
Update instructions on dev env
