### 2024-06-05 57a9a00
Fix instructions for dev env
