### 2024-06-06 27b13bd
make sure time diff type is timedelta: convert timedelta64 to timedelta in such case
