### 2024-06-06 56d2b7f
make sure time diff type is timedelta: convert timedelta64 to timedelta in such case
