### 2024-06-06 c5ecd89
Use `Union` for skipna to allow bool or None, instead of using `|` operator that fails CI build for python 3.9
