+++ xcdat/spatial.py
@@ -73,7 +73,7 @@ class SpatialAccessor:
         keep_weights: bool = False,
         lat_bounds: Optional[RegionAxisBounds] = None,
         lon_bounds: Optional[RegionAxisBounds] = None,
-        skipna: bool | None = None,
+        skipna: Union[bool, None] = None,
     ) -> xr.Dataset:
         """
         Calculates the spatial average for a rectilinear grid over an optionally
@@ -708,7 +708,7 @@ class SpatialAccessor:
         self,
         data_var: xr.DataArray,
         axis: List[SpatialAxis],
-        skipna: bool | None = None,
+        skipna: Union[bool, None] = None,
     ) -> xr.DataArray:
         """Perform a weighted average of a data variable.
 
