### 2024-06-06 826a7f1
Use `Union` for skipna to allow bool or None, instead of using `|` operator that fails CI build for python 3.9 (fixing those missed in the previous commit)
