### 2024-06-06 d2fdd98
Use `Union` for skipna to allow bool or None, instead of using `|` operator that fails CI build for python 3.9 (fixing those missed in the previous commit)
