+++ xcdat/bounds.py
@@ -640,10 +640,7 @@ class BoundsAccessor:
         elif freq == "hour":
             # Determine the daily frequency for generating time bounds.
             if daily_subfreq is None:
-                diff = pd.to_timedelta(time.values[1] - time.values[0])
-                # `cftime` objects only support arithmetic using `timedelta` objects, so
-                # the values of `diff` must be casted from `dtype="timedelta64[ns]"`
-                # to `timedelta` objects.
+                diff = time.values[1] - time.values[0]
                 hrs = diff.seconds / 3600
                 daily_subfreq = int(24 / hrs)  # type: ignore
 
