### 2024-06-06 69c1a29
make sure time diff type is timedelta: in case it was timedelta64 convert it, so following `.seconds` line could run without error
