### 2024-06-06 a60007f
Update xcdat/bounds.py
Co-authored-by: Stephen Po-Chedley <pochedley@gmail.com>
