### 2024-06-06 c019318
proceed `pd.to_timedelta` only if diff type is `np.timedelta64`
