### 2024-06-06 09203f2
Update `_create_time_bounds()` - Add comment about needing to convert dype=timedelta64[ns] to pandas timedelta object - Ignore C901 flake8 error
