+++ xcdat/bounds.py
@@ -504,7 +504,7 @@ class BoundsAccessor:
 
         return list(set(keys))
 
-    def _create_time_bounds(
+    def _create_time_bounds(  # noqa: C901
         self,
         time: xr.DataArray,
         freq: Optional[Literal["year", "month", "day", "hour"]] = None,
@@ -601,14 +601,16 @@ class BoundsAccessor:
         elif freq == "day":
             time_bnds = self._create_daily_time_bounds(timesteps, obj_type)
         elif freq == "hour":
-            # Determine the daily frequency for generating time  bounds.
+            # Determine the daily frequency for generating time bounds.
             if daily_subfreq is None:
                 diff = time.values[1] - time.values[0]
+
+                # Arrays with `dtype="timedelta64[ns]"` must be converted to
+                # pandas timedelta objects in order to access the `.seconds`
+                # time component.
                 if isinstance(diff, np.timedelta64):
                     diff = pd.to_timedelta(diff)
-                    # `cftime` objects only support arithmetic using `timedelta` objects, so
-                    # the values of  `diffs` must be casted from `dtype="timedelta64[ns]"`
-                    # to `timedelta` objects.
+
                 hrs = diff.seconds / 3600
                 daily_subfreq = int(24 / hrs)  # type: ignore
 
