### 2024-06-07 xCDAT/xcdat#661 @pochedls
@brian-rose – thanks for reporting this. We are currently [going through](https://github.com/xCDAT/xcdat/issues/589) all of our docs to catch issues like this; it looks like this [was caught](https://github.com/xCDAT/xcdat/blob/32bb27f738f2dfc1381830b9e66661dcc981f53b/docs/examples/introduction-to-xcdat.ipynb) and the [PR is being finalized](https://github.com/xCDAT/xcdat/pull/650).

We previously accepted `add_bounds` as a Boolean, which would add bounds to all axes. This was [deprecated](https://github.com/xCDAT/xcdat/commit/472c04b36aa89eb7f9d6d49cff0b0851dfb8edce#diff-e5b713e4bf50c61361496f77fdadcf9d3ffc2fcb39c4c953932cd780ae951e55L560) in favor of having the user specify the axes which bounds should be added to (if missing). We included deprecation warnings before the change, but obviously missed this in the documentation. We will fix this soon.
