### 2024-06-07 #6426 @brian-rose
@tomvothecoder and @arfon:

I have finished going through all the documentation and tested out the functionality. Specifically, these are some things I went through:
- Read through the whole [Getting Started](https://xcdat.readthedocs.io/en/latest/getting-started-guide/index.html) section of the docs
- Installed the package locally, following the installation instructions
- Read through and executed all the example notebooks in the [gallery](https://xcdat.readthedocs.io/en/latest/gallery.html)
- Skimmed through the [API Reference](https://xcdat.readthedocs.io/en/latest/api.html)
- Read the [Contributing Guide](https://xcdat.readthedocs.io/en/latest/contributing.html)
- Forked and cloned the source repository
- Installed the development environment following [these instructions](https://xcdat.readthedocs.io/en/latest/contributing.html#creating-a-development-environment)
- Ran the automated test suite with `make test`

Overall the docs are high quality. Almost every I looked at was clear and clean and easy to understand. This was a really valuable exercise for me, because I've learned that xCDAT is software that will really help me and my group do a lot routine data processing more consistently and expressively. Great package and great documentation!

The JOSS paper is also very well-written and easy to follow. It offers a clear and compelling statement of need. One of the central needs is easy access to parallelization for the out-of-core that are common in "big data" climate analysis. The benchmark calculations in Figure 1 of the paper illustrate this vividly. Ironically (and as noted above) I wasn't able to reproduce these calculations because I don't have access to a machine with large enough memory to handle the serial cases. Parallel is clearly the way to go.

My only hang-up right now is https://github.com/xCDAT/xcdat/issues/661. I suggest that the team prioritize resolving this, because it's the very first line of xCDAT code in the [very first tutorial in the docs](https://xcdat.readthedocs.io/en/latest/examples/introduction-to-xcdat.html) which is failing.
