### 2024-06-07 8879f8f
Update CF Metadata Convention references
