### 2024-06-07 504674a
Add info about lazily decoding cftime
