### 2024-06-07 xCDAT/xcdat#661 @tomvothecoder
@brian-rose Thanks for reporting this! I just merged PR #650 which addresses this issue. The "latest" version on Read The Docs will reflect this change soon: https://xcdat.readthedocs.io/en/latest/examples/introduction-to-xcdat.html.
