### 2024-06-07 #6426 @tomvothecoder
Thank you @brian-rose! I appreciate your diligent review and all of your helpful comments. I'm glad to hear that you had an overall great experience reviewing xCDAT :)

> This was a really valuable exercise for me, because I've learned that xCDAT is software that will really help me and my group do a lot routine data processing more consistently and expressively. Great package and great documentation!

If you ever need anything related to xCDAT, let me or Steve (@pochedls) know. We'd love to help you and your group utilize xCDAT in your data processing work wherever possible.

> My only hang-up right now is [xCDAT/xcdat#661](https://github.com/xCDAT/xcdat/issues/661), which is the reason I've left the "Example usage" box unchecked. I suggest that the team prioritize resolving this, because it's the very first line of xCDAT code in the [very first tutorial in the docs](https://xcdat.readthedocs.io/en/latest/examples/introduction-to-xcdat.html) which is failing.

I just addressed this issue and commented about it [here](https://github.com/xCDAT/xcdat/issues/661#issuecomment-2155495817). You should be able to re-run the notebook with this latest commit on `main`. Let me know if it works for you!
That should take care of the "Example usage" checkbox.
