### 2024-06-10 1302acb
Replace svg with png in index.rst to fix rtd build
