+++ docs/getting-started-guide/overview.rst
@@ -165,21 +165,12 @@ Xarray and Xarray usage in climate science!
 - `Pangeo Forum <https://foundations.projectpythia.org/core/xarray.html>`_
 - `Project Pythia <https://foundations.projectpythia.org/core/xarray.html>`_
 
-Projects Using xCDAT
---------------------
-
-xCDAT is actively being integrated as a core component of the `Program for Climate Model
-Diagnosis and Intercomparison (PCMDI) Metrics Package`_ and the `Energy Exascale Earth
-System Model Diagnostics (E3SM) Package`_. xCDAT is also included in the `E3SM Unified
-Anaconda Environment`_ that is deployed on various U.S. Department of Energy
-supercomputers to run E3SM software tools.
-
-.. _Program for Climate Model Diagnosis and Intercomparison (PCMDI) Metrics Package: https://pcmdi.github.io/pcmdi_metrics/
-.. _Energy Exascale Earth System Model Diagnostics (E3SM) Package: https://e3sm-project.github.io/e3sm_diags/_build/html/main/index.html
-.. _E3SM Unified Anaconda Environment: https://e3sm.org/resources/tools/other-tools/e3sm-unified-environment/
-
 Acknowledgement
 ---------------
+
+Huge thank you to all of the xCDAT `contributors`_!
+
+.. _contributors: https://github.com/xCDAT/xcdat/graphs/contributors
 
 xCDAT is jointly developed by scientists and developers from the Energy Exascale
 Earth System Model (`E3SM`_) Project and Program for Climate Model Diagnosis and
@@ -199,15 +190,6 @@ Environmental Research (`BER`_) within the `Department of Energy`_'s `Office of 
 .. _BER: https://science.osti.gov/ber
 .. _Department of Energy: https://www.energy.gov/
 .. _Office of Science: https://science.osti.gov/
-
-Contributors
-------------
-
-Thank you to all of our contributors!
-
-.. image:: https://contrib.rocks/image?repo=xCDAT/xcdat
-  :alt: xCDAT contributors
-  :target: https://github.com/xCDAT/xcdat/graphs/contributors
 
 License
 -------
