### 2024-06-10 b3ed34c
Remove contributors graph
