### 2024-06-10 c6b314a
Add coverage for new code
