### 2024-06-10 c653bf2
[PR] time type conversion to timedelta if timedelta64 is given (#660)
* make sure time diff type is timedelta: in case it was timedelta64 convert it, so following `.seconds` line could run without error

* Update xcdat/bounds.py

Co-authored-by: Stephen Po-Chedley <pochedley@gmail.com>

* proceed `pd.to_timedelta` only if diff type is `np.timedelta64`

* Update `_create_time_bounds()`
- Add comment about needing to convert dype=timedelta64[ns] to pandas timedelta object
- Ignore C901 flake8 error

* Add coverage for new code

* Update tests/test_bounds.py

---------

Co-authored-by: Stephen Po
[…]
