### 2024-06-12 #6426 @brian-rose
Confirming here that the [updated notebook](https://xcdat.readthedocs.io/en/latest/examples/introduction-to-xcdat.html) works as expected!
