### 2024-06-12 #6426 @brian-rose
I checked off the "Example usage" box, and I enthusiastically recommend this paper for publication in JOSS.
