+++ docs/demos/24-07-11-scipy-2024/wget_script_2024-6-12_13-17-40.sh
@@ -1,0 +1,1089 @@
+#!/bin/bash
+##############################################################################
+# ESG Federation download script
+#
+# Template version: 1.2
+# Generated by esgf-node.llnl.gov - 2024/06/12 13:17:40
+# Search URL: https://esgf-node.llnl.gov/esg-search/wget?dataset_id=E3SM.2_0.historical.LR.atmos.native.model-output.mon.ens1.v20220806|esgf-data2.llnl.gov
+#
+###############################################################################
+# first be sure it's bash... anything out of bash or sh will break
+# and the test will assure we are not using sh instead of bash
+if [ $BASH ] && [ `basename $BASH` != bash ]; then
+    echo "######## This is a bash script! ##############"
+    echo "Change the execution bit 'chmod u+x $0' or start with 'bash $0' instead of sh."
+    echo "Trying to recover automatically..."
+    sleep 1
+    /bin/bash $0 $@
+    exit $?
+fi
+
+version=1.3.2
+CACHE_FILE=.$(basename $0).status
+openId=
+search_url='https://esgf-node.llnl.gov/esg-search/wget?dataset_id=E3SM.2_0.historical.LR.atmos.native.model-output.mon.ens1.v20220806|esgf-data2.llnl.gov'
+
+#These are the embedded files to be downloaded
+download_files="$(cat <<EOF--dataset.file.url.chksum_type.chksum
+'v2.LR.historical_0101.eam.h0.1850-01.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/E3SM/2_0/historical/LR/atmos/native/model-output/mon/ens1/v20220806/v2.LR.historical_0101.eam.h0.1850-01.nc' 'SHA256' '81d46fbbd78902c9fbf6b245bab2a19eaf6a2797340bd8d89cce578d2451955b'
+'v2.LR.historical_0101.eam.h0.1850-02.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/E3SM/2_0/historical/LR/atmos/native/model-output/mon/ens1/v20220806/v2.LR.historical_0101.eam.h0.1850-02.nc' 'SHA256' 'ac4bd6778dce9b872173609639f511e0d963ddc7fef09ea5c9cd52af97987b90'
+'v2.LR.historical_0101.eam.h0.1850-03.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/E3SM/2_0/historical/LR/atmos/native/model-output/mon/ens1/v20220806/v2.LR.historical_0101.eam.h0.1850-03.nc' 'SHA256' '395224440bc77694b0a41254d0a42f6b68e113584a3d40e5a04b1107c23b710a'
+'v2.LR.historical_0101.eam.h0.1850-04.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/E3SM/2_0/historical/LR/atmos/native/model-output/mon/ens1/v20220806/v2.LR.historical_0101.eam.h0.1850-04.nc' 'SHA256' 'd8a98682c289f9c58e46dea061de366eda16967d667526e1aa503141e478e738'
[…]
