+++ docs/paper/paper.md
@@ -27,7 +27,7 @@ authors:
 affiliations:
     - name: Lawrence Livermore National Lab, Livermore, USA
       index: 1
-date: 23 January 2024
+date: 24 July 2024
 bibliography: paper.bib
 ---
 
