### 2024-06-24 1b03e61
Add JOSS paper (#567)
