+++ docs/paper/paper.md
@@ -1,0 +1,108 @@
+---
+title: "xCDAT: A Python Package for Simple and Robust Analysis of Climate Data"
+tags:
+    - Python
+    - python
+    - xarray
+    - climate science
+    - climate research
+    - climate data
+    - climate data analysis
+authors:
+    - name: Tom Vo
+      orcid: 0000-0002-2461-0191
+      affiliation: 1
+    - name: Stephen Po-Chedley
+      orcid: 0000-0002-0390-238X
+      affiliation: 1
+    - name: Jason Boutte
+      orcid: 0009-0009-3996-3772
+      affiliation: 1
+    - name: Jiwoo Lee
+      orcid: 0000-0002-0016-7199
+      affiliation: 1
+    - name: Chengzhu Zhang
+      orcid: 0000-0002-9632-0716
+      affiliation: 1
+affiliations:
+    - name: Lawrence Livermore National Lab, Livermore, USA
+      index: 1
+date: 24 July 2024
+bibliography: paper.bib
+---
+
+# Summary
+
+xCDAT (Xarray Climate Data Analysis Tools) is an open-source Python package that extends Xarray [@Hoyer:2017] for climate data analysis on structured grids. xCDAT streamlines analysis of climate data by exposing common climate analysis operations through a set of straightforward APIs. Some of xCDAT's key features include spatial averaging, temporal averaging, and regridding. These features are inspired by the Community Data Analysis Tools ([CDAT](https://cdat.llnl.gov/)) library [@Williams:2009] [@Williams:2014] [@cdat] and leverage powerful packages in the [Xarray](https://docs.xarray.dev/en/stable/) ecosystem including [xESMF](https://github.com/pangeo-data/xESMF) [@xesmf], [xgcm](https://xgcm.readthedocs.io/en/latest/) [@xgcm], and [CF xarray](https://cf-xarray.readthedocs.io/en/latest/) [@cf-xarray]. To ensure general compatibility across various climate models, xCDAT operates on datasets that are compliant with the Climate and Forecast (CF) metadata conventions [@Hassell:2017].
+
+# Statement of Need
+
+Analysis of climate data frequently requires a number of core operations, including reading and writing of netCDF files, horizontal and vertical regridding, and spatial and temporal averaging. While many individual software packages address these needs in a variety of computational languages, CDAT stands out because it provides these essential operations via open-source, interoperable Python packages. Since CDAT’s inception, the volume of climate data has grown substantially as a result of both a larger pool of data products and increasing spatiotemporal resolution of model and observational data. Larger data stores are important for advancing geoscientific understanding, but also require increasingly performant software and hardware. These factors have sparked a need for new analysis software that offers core geospatial analysis functionalities capable of efficiently handling large datasets while using modern technologies and standardized software engineering principles.
+
+xCDAT addresses this need by combining the power of Xarray with meticulously developed geospatial analysis features inspired by CDAT. Xarray is the foundation of xCDAT because of its widespread adoption, technological maturity, and ability to handle large datasets with parallel computing via Dask. Xarray is also interoperable with the scientific Python ecosystem (e.g., [NumPy](https://numpy.org/), [pandas](https://pandas.pydata.org/), [Matplotlib](https://matplotlib.org/)), which greatly benefits users who need to use additional analysis tools. Since Xarray is designed as a general-purpose library, xCDAT fills in domain specific gaps by providing features to serve the climate science community _(refer to [Key Features](#key-features))_.
+
[…]
