### 2024-06-24 b244601
Update changelog and fix date in `paper.md`
