### 2024-06-24 dc49144
Apply suggestions from code review
