### 2024-06-24 39bb46c
Update docs/paper/paper.md
