### 2024-06-24 6a4b721
Constrain `numpy<2.0`
