+++ conda-env/dev.yml
@@ -11,7 +11,7 @@ dependencies:
     - cftime
     - dask
     - netcdf4
-    - numpy >=1.23.0
+    - numpy >=1.23.0,<2.0
     - pandas
     - python-dateutil
     - xarray >=2022.02.0
