+++ .github/workflows/build_workflow.yml
@@ -95,7 +95,8 @@ jobs:
               run: |
                   mamba env update -n xcdat_ci -f conda-env/ci.yml
                   # Make sure the Python version in the env matches the current matrix version.
-                  mamba install -c conda-forge python=${{ matrix.python-version }}
+                  # Make sure numpy is not > 2.0.
+                  mamba install -c conda-forge python=${{ matrix.python-version }} "numpy>=1.23.0,<2.0"
 
             - name: Install xcdat
               # Source: https://github.com/conda/conda-build/issues/4251#issuecomment-1053460542
