### 2024-06-24 #6426 @tomvothecoder
Hi @arfon, I completed your final checklist your [comment](https://github.com/openjournals/joss-reviews/issues/6426#issuecomment-2184005026) above. Thanks!

- [x] please give your own paper a final read to check for any potential typos etc. -- I only updated the date to 24 June 2024 [here](https://github.com/xCDAT/xcdat/blob/main/docs/paper/paper.md)
- [x] make a new release of this software that includes the changes that have resulted from this review (link)(https://github.com/xCDAT/xcdat/releases/tag/v0.7.1)
- [x] make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive ([link](https://zenodo.org/records/12522560))
  - [x] The title of the archive is the same as the JOSS paper title
  - [x] That the authors of the archive are the same as the JOSS paper authors
