### 2024-06-25 de29e5b
Update notebook and add wget scripts
