+++ docs/demos/24-07-11-scipy-2024/ta_wget_script_2024-6-25_14-15-11.sh
@@ -3,14 +3,14 @@
 # ESG Federation download script
 #
 # Template version: 1.2
-# Generated by esgf-node.llnl.gov - 2024/06/12 13:17:40
-# Search URL: https://esgf-node.llnl.gov/esg-search/wget?dataset_id=E3SM.2_0.historical.LR.atmos.native.model-output.mon.ens1.v20220806|esgf-data2.llnl.gov
+# Generated by esgf-node.llnl.gov - 2024/06/25 14:15:11
+# Search URL: https://esgf-node.llnl.gov/esg-search/wget?dataset_id=CMIP6.CMIP.E3SM-Project.E3SM-2-0.historical.r1i1p1f1.day.tas.gr.v20220923|esgf-data2.llnl.gov
 #
 ###############################################################################
 # first be sure it's bash... anything out of bash or sh will break
 # and the test will assure we are not using sh instead of bash
 if [ $BASH ] && [ `basename $BASH` != bash ]; then
-    echo "######## This is a bash script! ##############"
+    echo "######## This is a bash script! ##############" 
     echo "Change the execution bit 'chmod u+x $0' or start with 'bash $0' instead of sh."
     echo "Trying to recover automatically..."
     sleep 1
@@ -21,58 +21,27 @@ fi
 version=1.3.2
 CACHE_FILE=.$(basename $0).status
 openId=
-search_url='https://esgf-node.llnl.gov/esg-search/wget?dataset_id=E3SM.2_0.historical.LR.atmos.native.model-output.mon.ens1.v20220806|esgf-data2.llnl.gov'
+search_url='https://esgf-node.llnl.gov/esg-search/wget?dataset_id=CMIP6.CMIP.E3SM-Project.E3SM-2-0.historical.r1i1p1f1.day.tas.gr.v20220923|esgf-data2.llnl.gov'
 
 #These are the embedded files to be downloaded
 download_files="$(cat <<EOF--dataset.file.url.chksum_type.chksum
-'v2.LR.historical_0101.eam.h0.1850-01.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/E3SM/2_0/historical/LR/atmos/native/model-output/mon/ens1/v20220806/v2.LR.historical_0101.eam.h0.1850-01.nc' 'SHA256' '81d46fbbd78902c9fbf6b245bab2a19eaf6a2797340bd8d89cce578d2451955b'
-'v2.LR.historical_0101.eam.h0.1850-02.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/E3SM/2_0/historical/LR/atmos/native/model-output/mon/ens1/v20220806/v2.LR.historical_0101.eam.h0.1850-02.nc' 'SHA256' 'ac4bd6778dce9b872173609639f511e0d963ddc7fef09ea5c9cd52af97987b90'
[…]
