### 2024-06-25 edf3ba8
Fix wget scripts and add latest notebook changes
