+++ docs/demos/24-07-11-scipy-2024/cl_Amon_E3SM-2-0_historical_r1i1p1f1_gr/cl_Amon_wget_script_2024-6-25_14-39-4.sh
@@ -3,14 +3,14 @@
 # ESG Federation download script
 #
 # Template version: 1.2
-# Generated by esgf-node.llnl.gov - 2024/06/25 14:15:11
-# Search URL: https://esgf-node.llnl.gov/esg-search/wget?dataset_id=CMIP6.CMIP.E3SM-Project.E3SM-2-0.historical.r1i1p1f1.day.tas.gr.v20220923|esgf-data2.llnl.gov
+# Generated by esgf-node.llnl.gov - 2024/06/25 14:39:04
+# Search URL: https://esgf-node.llnl.gov/esg-search/wget?dataset_id=CMIP6.CMIP.E3SM-Project.E3SM-2-0.historical.r1i1p1f1.Amon.cl.gr.v20220830|esgf-data2.llnl.gov
 #
 ###############################################################################
 # first be sure it's bash... anything out of bash or sh will break
 # and the test will assure we are not using sh instead of bash
 if [ $BASH ] && [ `basename $BASH` != bash ]; then
-    echo "######## This is a bash script! ##############" 
+    echo "######## This is a bash script! ##############"
     echo "Change the execution bit 'chmod u+x $0' or start with 'bash $0' instead of sh."
     echo "Trying to recover automatically..."
     sleep 1
@@ -21,27 +21,14 @@ fi
 version=1.3.2
 CACHE_FILE=.$(basename $0).status
 openId=
-search_url='https://esgf-node.llnl.gov/esg-search/wget?dataset_id=CMIP6.CMIP.E3SM-Project.E3SM-2-0.historical.r1i1p1f1.day.tas.gr.v20220923|esgf-data2.llnl.gov'
+search_url='https://esgf-node.llnl.gov/esg-search/wget?dataset_id=CMIP6.CMIP.E3SM-Project.E3SM-2-0.historical.r1i1p1f1.Amon.cl.gr.v20220830|esgf-data2.llnl.gov'
 
 #These are the embedded files to be downloaded
 download_files="$(cat <<EOF--dataset.file.url.chksum_type.chksum
-'tas_day_E3SM-2-0_historical_r1i1p1f1_gr_18500101-18591231.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/CMIP6/CMIP/E3SM-Project/E3SM-2-0/historical/r1i1p1f1/day/tas/gr/v20220923/tas_day_E3SM-2-0_historical_r1i1p1f1_gr_18500101-18591231.nc' 'SHA256' '1e725632935b2281c61ea9f4088af6ebdb51b91d94734e82495c846bb0b6b8ca'
-'tas_day_E3SM-2-0_historical_r1i1p1f1_gr_18600101-18691231.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/CMIP6/CMIP/E3SM-Project/E3SM-2-0/historical/r1i1p1f1/day/tas/gr/v20220923/tas_day_E3SM-2-0_historical_r1i1p1f1_gr_18600101-18691231.nc' 'SHA256' '88c2e873e45d69fdd9c90ce5e1243895dd88d951d9670e0724c4db169f9c4913'
-'tas_day_E3SM-2-0_historical_r1i1p1f1_gr_18700101-18791231.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/CMIP6/CMIP/E3SM-Project/E3SM-2-0/historical/r1i1p1f1/day/tas/gr/v20220923/tas_day_E3SM-2-0_historical_r1i1p1f1_gr_18700101-18791231.nc' 'SHA256' '59bcf4c4c196be78b30dfbad66507aaf88f9fabd82f1574ff288c7bf99d67650'
[…]
