### 2024-06-27 53674fc
Add instructions for getting esgf data with wget script
