+++ docs/demos/24-07-11-scipy-2024/cl_Amon_E3SM-2-0_historical_r1i1p1f1_gr/cl_Amon_wget_script_2024-6-25_14-39-4.sh
@@ -1,1046 +1,0 @@
-#!/bin/bash
-##############################################################################
-# ESG Federation download script
-#
-# Template version: 1.2
-# Generated by esgf-node.llnl.gov - 2024/06/25 14:39:04
-# Search URL: https://esgf-node.llnl.gov/esg-search/wget?dataset_id=CMIP6.CMIP.E3SM-Project.E3SM-2-0.historical.r1i1p1f1.Amon.cl.gr.v20220830|esgf-data2.llnl.gov
-#
-###############################################################################
-# first be sure it's bash... anything out of bash or sh will break
-# and the test will assure we are not using sh instead of bash
-if [ $BASH ] && [ `basename $BASH` != bash ]; then
-    echo "######## This is a bash script! ##############"
-    echo "Change the execution bit 'chmod u+x $0' or start with 'bash $0' instead of sh."
-    echo "Trying to recover automatically..."
-    sleep 1
-    /bin/bash $0 $@
-    exit $?
-fi
-
-version=1.3.2
-CACHE_FILE=.$(basename $0).status
-openId=
-search_url='https://esgf-node.llnl.gov/esg-search/wget?dataset_id=CMIP6.CMIP.E3SM-Project.E3SM-2-0.historical.r1i1p1f1.Amon.cl.gr.v20220830|esgf-data2.llnl.gov'
-
-#These are the embedded files to be downloaded
-download_files="$(cat <<EOF--dataset.file.url.chksum_type.chksum
-'cl_Amon_E3SM-2-0_historical_r1i1p1f1_gr_185001-189912.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/CMIP6/CMIP/E3SM-Project/E3SM-2-0/historical/r1i1p1f1/Amon/cl/gr/v20220830/cl_Amon_E3SM-2-0_historical_r1i1p1f1_gr_185001-189912.nc' 'SHA256' '8af00a01428809a741bb676c5310e82d84857d16903ea28d0e938a80f2bb6fe5'
-'cl_Amon_E3SM-2-0_historical_r1i1p1f1_gr_190001-194912.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/CMIP6/CMIP/E3SM-Project/E3SM-2-0/historical/r1i1p1f1/Amon/cl/gr/v20220830/cl_Amon_E3SM-2-0_historical_r1i1p1f1_gr_190001-194912.nc' 'SHA256' 'a7063946dd80e7b75f37bb9402aeefb261ffba586c70f2511a7e1c49855f8ed8'
-'cl_Amon_E3SM-2-0_historical_r1i1p1f1_gr_195001-199912.nc' 'https://esgf-data2.llnl.gov/thredds/fileServer/user_pub_work/CMIP6/CMIP/E3SM-Project/E3SM-2-0/historical/r1i1p1f1/Amon/cl/gr/v20220830/cl_Amon_E3SM-2-0_historical_r1i1p1f1_gr_195001-199912.nc' 'SHA256' 'c4969d3b2d615f4023b55e70e82c986e1f78231a0c9835145091401e32fd3013'
[…]
