### 2024-06-28 01f0b6e
initial attempt at #531 (for spatial averaging)
