### 2024-06-28 48853f4
Add weight threshold option for spatial averaging
- Add parameter `min_weight` to `SpatialAccessor.average()`
