+++ tests/test_spatial.py
@@ -15,13 +15,13 @@ class TestSpatialAccessor:
         )
 
     def test__init__(self):
-        ds = self.ds.copy()
+        ds = self.ds.copy(deep=True)
         obj = SpatialAccessor(ds)
 
         assert obj._dataset.identical(ds)
 
     def test_decorator_call(self):
-        ds = self.ds.copy()
+        ds = self.ds.copy(deep=True)
         obj = ds.spatial
 
         assert obj._dataset.identical(ds)
@@ -50,7 +50,7 @@ class TestAverage:
             self.ds.spatial.average("ts", axis=["Y", "incorrect_axis"])
 
     def test_raises_error_if_lat_axis_coords_cant_be_found(self):
-        ds = self.ds.copy()
+        ds = self.ds.copy(deep=True)
 
         # Update CF metadata to invalid values so cf_xarray can't interpret them.
         del ds.lat.attrs["axis"]
@@ -64,7 +64,7 @@ class TestAverage:
             ds.spatial.average("ts", axis=["X", "Y"])
 
     def test_raises_error_if_lon_axis_coords_cant_be_found(self):
-        ds = self.ds.copy()
+        ds = self.ds.copy(deep=True)
 
         # Update CF metadata to invalid values so cf_xarray can't interpret them.
         del ds.lon.attrs["axis"]
@@ -141,13 +141,13 @@ class TestAverage:
             self.ds.spatial.average("ts", axis=["X", "Y"], weights=weights)
 
     def test_spatial_average_for_lat_region_and_keep_weights(self):
-        ds = self.ds.copy()
+        ds = self.ds.copy(deep=True)
 
         result = ds.spatial.average(
             "ts", axis=["Y"], lat_bounds=(-5.0, 5), keep_weights=True
         )
 
-        expected = self.ds.copy()
+        expected = self.ds.copy(deep=True)
         expected["ts"] = xr.DataArray(
             data=np.array(
                 [[2.25, 2.25, 2.25, 2.25], [1.0, 1.0, 1.0, 1.0], [1.0, 1.0, 1.0, 1.0]]
@@ -164,12 +164,12 @@ class TestAverage:
         xr.testing.assert_allclose(result, expected)
 
     def test_spatial_average_for_lat_region(self):
-        ds = self.ds.copy()
+        ds = self.ds.copy(deep=True)
 
         # Specifying axis as a str instead of list of str.
         result = ds.spatial.average("ts", axis=["Y"], lat_bounds=(-5.0, 5))
 
-        expected = self.ds.copy()
+        expected = self.ds.copy(deep=True)
         expected["ts"] = xr.DataArray(
             data=np.array(
                 [[2.25, 2.25, 2.25, 2.25], [1.0, 1.0, 1.0, 1.0], [1.0, 1.0, 1.0, 1.0]]
@@ -205,7 +205,7 @@ class TestAverage:
     def test_spatial_average_for_domain_wrapping_p_meridian_non_cf_conventions(
         self,
     ):
-        ds = self.ds.copy()
+        ds = self.ds.copy(deep=True)
 
         # get spatial average for original dataset
         ref = ds.spatial.average("ts").ts
@@ -222,14 +222,14 @@ class TestAverage:
 
     @requires_dask
     def test_spatial_average_for_lat_region_and_keep_weights_with_dask(self):
-        ds = self.ds.copy().chunk(2)
+        ds = self.ds.copy(deep=True).chunk(2)
 
         # Specifying axis as a str instead of list of str.
         result = ds.spatial.average(
             "ts", axis=["Y"], lat_bounds=(-5.0, 5), keep_weights=True
         )
 
[…]
