### 2024-06-28 1b4742c
initial attempt at #531 (for spatial averaging)
