### 2024-06-28 17458a8
cleanup print statement and complete code coverage
