### 2024-06-28 f69853e
cleanup print statement and complete code coverage
