### 2024-06-28 9967127
cleanup print statement and complete code coverage
