### 2024-07-01 #6426 @tomvothecoder
> @brian-rose, @mgrover1 – many thanks for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
>
> @tomvothecoder – your paper is now accepted and published in JOSS ⚡🚀💥

Thanks @arfon for facilitating the review process for this paper! I am grateful and excited to have this xCDAT paper published in JOSS. This is a great milestone for the project.

And big thanks to @brian-rose and @mgrover1 for your diligent review. I appreciate your time and interest in xCDAT! Feel free to reach out if you ever need anything xCDAT related (or a review of a paper like this) :)
